% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_aandeel.R
\docType{class}
\name{aandeel-class}
\alias{aandeel-class}
\title{S4-klasse die aandeel grondvlak of volume van bepaalde boomsoort berekend}
\description{
Deze klasse Aandeel staat in voor de berekening van waarden voor TypeVariabele Aandeel op basis van opgegeven kenmerken.  Ze is een nakomeling van de klasse AnalyseVariabele.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

