% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_aandeelKruidlaag.R
\docType{class}
\name{aandeelKruidlaag-class}
\alias{aandeelKruidlaag-class}
\title{S4-klasse die aandeel bedekking soorten binnen de kruidlaag berekend}
\description{
Deze klasse AandeelKruidlaag staat in voor de berekening van waarden voor
TypeVariabele AandeelKruidlaag op basis van opgegeven kenmerken.  Ze is een
nakomeling van de klasse bedekking.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk,
TypeKenmerk, WaardeMin en WaardeMax}
}}

