% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_aantal.R
\docType{class}
\name{aantal-class}
\alias{aantal-class}
\title{S4-klasse die aantal soorten met een bepaalde bedekking berekent}
\description{
Deze klasse Aantal staat in voor de berekening van waarden voor TypeVariabele Aantal op basis van opgegeven kenmerken.  Ze is een nakomeling van de klasse AnalyseVariabele.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

