% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_aantalGroepen.R
\docType{class}
\name{aantalGroepen-class}
\alias{aantalGroepen-class}
\title{S4-klasse die het aantal groepen telt}
\description{
Deze klasse aantalGroepen staat in voor de berekening van waarden voor
AnalyseVariabele aantalGroepen op basis van opgegeven kenmerken.  Ze is een
nakomeling van de klasse aantal.  Ze telt het aantal studieitems dat
opgegeven is, en als deze niet opgegeven zijn, bepaalt ze eerst waarden voor
de studieitems op basis van de soortenlijst.  Het gaat hier over
studiegroepen als levensvormen of groeivormen, en voorbeelden van studieitems
zijn helofyten of dwergstruiken.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden
Vegetatielaag, Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

