% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_AnalyseVariabele.R
\docType{class}
\name{AnalyseVariabele-class}
\alias{AnalyseVariabele-class}
\title{S4-klasse waarin de generieke}
\description{
Deze virtuele klasse Waarde geeft de mogelijkheid om verschillende typen waarden (getallen, categorische variabelen,...) in eenzelfde veld te bewaren en er gelijkbaardige bewerkingen op uit te voeren (bv. met elkaar te vergelijken).
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

