% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_bedekking.R
\docType{class}
\name{bedekking-class}
\alias{bedekking-class}
\title{S4-klasse die de totale bedekking van de soorten berekent}
\description{
Deze klasse Bedekking staat in voor de berekening van waarden voor TypeVariabele Bedekking op basis van opgegeven kenmerken.  Ze is een nakomeling van de klasse AnalyseVariabele.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

