% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_bedekkingExcl.R
\docType{class}
\name{bedekkingExcl-class}
\alias{bedekkingExcl-class}
\title{S4-klasse die de totale bedekking van de soorten berekent die niet in de
lijst staan}
\description{
Deze klasse BedekkingExcl staat in voor de berekening van waarden voor
AnalyseVariabele BedekkingExcl op basis van opgegeven kenmerken.  Ze is een
nakomeling van de klasse AnalyseVariabele.  Ze berekent de totale bedekking
van soorten die niet in de opgegeven soortenlijst staan (dus alle soorten
behalve soorten uit soorten uit lijst).
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden
Vegetatielaag, Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

