% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_bedekkingLaag.R
\docType{class}
\name{bedekkingLaag-class}
\alias{bedekkingLaag-class}
\title{S4-klasse die de totale bedekking van een vegetatielaag berekent}
\description{
Deze klasse bedekkingLaag staat in voor de berekening van waarden voor AnalyseVariabele bedekkingLaag op basis van opgegeven kenmerken.  Ze is een nakomeling van de klasse bedekking.  Ze maakt de berekening op basis van de studiegroep als deze aanwezig is in de opname, en anders op basis van de soortengroep.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Vegetatielaag, Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

