% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_bedekkingLaagPlus.R
\docType{class}
\name{bedekkingLaagPlus-class}
\alias{bedekkingLaagPlus-class}
\title{S4-klasse die de totale bedekking van een vegetatielaag en taxongroep
berekent}
\description{
Deze klasse bedekkingLaagPlus staat in voor de berekening van waarden voor
AnalyseVariabele bedekkingLaagPlus op basis van opgegeven kenmerken.  Ze is
een nakomeling van de klasse bedekkingLaag.  Verschillend van bedekkingLaag
wordt ze gebruikt als de totale bedekking van een vegetatielaag en een
taxongroep berekend moet worden, bv. de totale bedekking van de moslaag en
klimop.  Bij het samenvoegen van de bedekkingen maakt ze gebruik van de
formule van Ficher (die rekening houdt met een gedeeltelijke overlap).
De berekening van de bedekking van de vegetatielaag maakt ze, zoals de
klasse bedekkingLaag, op basis van de studiegroep als deze aanwezig is in de
opname, en anders op basis van de soortengroep.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden
Vegetatielaag, Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

