% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenAantalSoorten.R
\name{berekenAantalSoorten}
\alias{berekenAantalSoorten}
\title{Berekent het aantal soorten van een bepaalde soortengroep die in een opname aanwezig is}
\usage{
berekenAantalSoorten(Data_soorten, Soortengroeplijst,
  Minimumniveau = "aanwezig", ConnectieLSVIhabitats = connecteerMetLSVIdb())
}
\arguments{
\item{Data_soorten}{Bedekkingen van de sleutelsoorten in de vorm van een data.frame met velden ID, Soort_NL of Soort_Latijn en Tansley (bedekking in Tansley-schaal).  Als een soortenlijst gegeven wordt die mogelijk voor de berekening van de LSVI gegroepeerd moet worden tot een hoger taxonomisch of morfologisch niveau, wordt best ook Percentage (bedekking in procent) gegeven.  Als dit niet gegeven wordt, zal deze groepering gebeuren op basis van de klassemiddens van de Tansley-schaal.}

\item{Soortengroeplijst}{dataframe waarin per niveau aangegeven wordt (tabel Niveau met een int die het niveau aangeeft) welke soortengroepen geselecteerd moeten worden (tabel SoortengroepIDs met een string waarin de ID's na mekaar weergegeven worden, gescheiden door een komma)}

\item{Minimumniveau}{Minimum bedekking voor een soort om meegeteld te worden.  Standaard worden alle aanwezige soorten meegeteld, maar er kan bv. ook gekozen worden om enkel soorten te tellen die ze minstens frequent of abundant aanwezig zijn.  De waarde 'afwezig' telt enkel de afwezige soorten.}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor de LSVI van habitats, in te stellen d.m.v. functie connecteerMetLSVIdb.}
}
\value{
Deze functie genereert de resultaten in de vorm van een tabel met voor elk ID en elke SoortenlijstID het aantal soorten.
}
\description{
Deze hulpfunctie bepaalt het aantal soorten van een bepaalde soortengroep dat aanwezig is in een bepaalde opname.  Alternatief kunnen ook het aantal soorten bepaald worden dat minstens frequent of abundant aanwezig is, of elke andere mogelijke categorie uit de Tansley-schaal (aan te geven in parameter Minimumniveau).  Omdat het een technische hulpfunctie is, wordt hier gebruik gemaakt van ID's uit de databank.  Beter is om gebruik te maken van de generieke functie berekenAnalyseVariabele waarbij het eenvoudiger is om de soortengroep te selecteren.
}
\examples{
library(readr)
Data_soorten <- 
    read_csv2(system.file("vbdata/opname_4010_gelayout_soorten.csv", package = "LSVI"))
Schaalomzetting <-
    read_csv2(system.file("schaaltabellen/Schaalomzetting_ToonS.csv", package = "LSVI"))
Data_soorten <- merge(Data_soorten, Schaalomzetting, 
                      by.x = "Bedekking", by.y = "Schaal_opname")
Soortengroeplijst <- "369,143"
berekenAantalSoorten(Data_soorten, Soortengroeplijst)
berekenAantalSoorten(Data_soorten, Soortengroeplijst, 
                     Minimumniveau = "Frequent")

}
