% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenAnalyseVariabele.R
\name{berekenAnalyseVariabele}
\alias{berekenAnalyseVariabele}
\title{Berekent de opgegeven AnalyseVariabele van de opgegeven soortgroep voor de opname}
\usage{
berekenAnalyseVariabele(ConnectieLSVIhabitats, AnalyseVariabele, Data_soorten,
  Soortengroeplijst)
}
\arguments{
\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor de LSVI van habitats, in te stellen d.m.v. functie connecteerMetLSVIdb.}

\item{AnalyseVariabele}{Zie description voor voorbeelden.}

\item{Data_soorten}{Bedekkingen van de sleutelsoorten in de vorm van een data.frame met velden ID, Soort_NL of Soort_Latijn en Percentage (bedekking in procent) of Tansley (bedekking in Tansley-schaal), afhankelijk van het type AnalyseVariabele.  Voor een analysevariabele startend met 'aantal' moet Data_soorten het veld Tansley bevatten (en bij voorkeur ook Percentage als een soortenlijst gegeven wordt die voor de LSVI-berekening gegroepeerd moet worden tot een hoger niveau, bv. Genusniveau), voor een analysevariabele startend met 'bedekking' het veld Bedekking.}

\item{Soortengroeplijst}{string waarin de SoortengroepID's na elkaar weergegeven worden, gescheiden door een komma}
}
\value{
Deze functie genereert de resultaten in de vorm van een tabel met voor elk ID en elke SoortenlijstID het aantal soorten of de bedekking van de soorten.
}
\description{
Deze functie bepaalt/berekent de opgegeven analysevariabele op basis van een opgegeven soortengroep voor de opgegeven opname.  Voorbeelden van analysevariabelen zijn 

- 'aantal_aanwezig' die het aantal soorten uit de opgegeven soortengroep telt die aanwezig zijn in de opname.

- 'aantal_frequent_aanwezig' die het aantal soorten uit de opgegeven soortengroep telt die minstens frequent aanwezig zijn in de opname.  (Analoge analysevariabelen bestaan voor alle categorieen uit de Tansley-schaal.)

- 'aantal_afwezig' die het aantal soorten uit de opgegeven soortengroep telt die niet aanwezig zijn in de opname, wat wil zeggen dat ze niet voorkomen in de opname of dat in de opname hun Tansley-bedekking als 'afwezig' gequoteerd is.

- 'bedekking_vegetatie' berekent voor de opname de totale bedekking van alle soorten uit de opgegeven soortengroep (in %).

- 'bedekking_vegetatie_Tansley' berekent voor de opname de totale bedekking van alle soorten uit de opgegeven soortengroep en zet deze om naar de Tansley-schaal

De berekening van andere analysevariabelen uit de databank ('grootte_grootste_vegetatievlek', 'aantal_aanwezig_kruidlaag' en 'bedekking_grondvlak') is sterk afhankelijk van de verzamelde gegevens en wordt daarom niet in een generieke functie gestoken.  Ze gebeurt voor elk van de verzamelde gegevens in de specifieke functie die op maat van de gegevens geschreven is.
}
\examples{
library(readr)
Data_soorten <- 
    read_csv2(system.file("vbdata/opname_4010_gelayout_soorten.csv", 
              package = "LSVI"))
Schaalomzetting <-
    read_csv2(system.file("schaaltabellen/Schaalomzetting_ToonS.csv", 
              package = "LSVI"))
Data_soorten <- merge(Data_soorten, Schaalomzetting, 
                      by.x = "Bedekking", by.y = "Schaal_opname")
                      
ConnectieLSVIhabitats <- connecteerMetLSVIdb()
berekenAnalyseVariabele(ConnectieLSVIhabitats, "aantal_aanwezig", 
                        Data_soorten, "369,143")
berekenAnalyseVariabele(ConnectieLSVIhabitats, "aantal_frequent_aanwezig", 
                        Data_soorten, "369,143")
berekenAnalyseVariabele(ConnectieLSVIhabitats, "bedekking_vegetatie", 
                        Data_soorten, "369,143")
berekenAnalyseVariabele(ConnectieLSVIhabitats, "bedekking_vegetatie_Tansley", 
                        Data_soorten, "369,143")
library(RODBC)
odbcClose(ConnectieLSVIhabitats)

}

