% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenBedekkingSoorten.R
\name{berekenBedekkingSoorten}
\alias{berekenBedekkingSoorten}
\title{Berekent de totale bedekking van de soorten van een bepaalde soortengroep die in een opname aanwezig is}
\usage{
berekenBedekkingSoorten(Data_soorten, Soortengroeplijst)
}
\arguments{
\item{Data_soorten}{Bedekkingen van de sleutelsoorten in de vorm van een data.frame met velden ID, Soort_NL of Soort_Latijn en Percentage (integer die de bedekking in % aangeeft). (Eventueel zou hier ook de NBNTaxonVersionKey kunnen gebruikt worden.)}

\item{Soortengroeplijst}{dataframe waarin per niveau aangegeven wordt (tabel Niveau met een int die het niveau aangeeft) welke soortengroepen geselecteerd moeten worden (tabel SoortengroepIDs met een string waarin de ID's na mekaar weergegeven worden, gescheiden door een komma)}
}
\value{
Deze functie genereert de resultaten in de vorm van een tabel met voor elk ID en elke SoortenlijstID de totale bedekking van de soorten.
}
\description{
Deze hulpfunctie bepaalt de bedekking van de soorten van een bepaalde soortengroep dat aanwezig is in een bepaalde opname.  Omdat het een technische hulpfunctie is, wordt hier gebruik gemaakt van ID's uit de databank.  Beter is om gebruik te maken van de generieke functie berekenAnalyseVariabele die meer mogelijkheden heeft voor het selecteren van de soortengroep.
}
\examples{
Data_soorten <- 
    read.csv2(system.file("vbdata/opname_4010_gelayout_soorten.csv", package = "LSVI"), 
              stringsAsFactors = FALSE)
Schaalomzetting <-
    read.csv2(system.file("schaaltabellen/Schaalomzetting_ToonS.csv", package = "LSVI"),
             stringsAsFactors = FALSE)
Data_soorten <- merge(Data_soorten, Schaalomzetting, 
                      by.x = "Bedekking", by.y = "Schaal_opname")
Soortengroeplijst <- "369,143"
berekenBedekkingSoorten(Data_soorten, Soortengroeplijst)

}

