% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenLSVI.R
\name{berekenLSVI}
\alias{berekenLSVI}
\title{Berekent de LSVI op basis van de opgegeven opnames}
\usage{
berekenLSVI(Versie = geefUniekeWaarden("Versie", "VersieLSVI"),
  Kwaliteitsniveau = c("alle", "1", "2"), Server = "inbosql03\\\\prd",
  Databank = "D0122_00_LSVIHabitatTypes", Gebruiker = "D0122_AppR",
  Wachtwoord = "***REMOVED***")
}
\arguments{
\item{Versie}{De versie van het LSVI-rapport op basis waarvan de berekening gemaakt wordt, bv. "Versie 2" of "Versie 3".  Bij de default "alle" wordt de LSVI volgens de verschillende versies berekend.}

\item{Kwaliteitsniveau}{Voor elke versie van de LSVI zijn er een of meerdere kwaliteitsniveaus gedefinieerd in de databank.  Zo is er bij Versie 2.0 een onderscheid gemaakt tussen goede staat (A), voldoende staat (B) en gedegradeerde staat (C).  Hier duidt kwaliteitsniveau 1 de grens tussen voldoende (B) en gedegradeerd (C) aan en kwaliteitsniveau 2 het onderscheid tussen goed (A) en voldoende (B).  Bij Versie 3 duidt kwaliteitsniveau 1 op het onderscheid tussen ongunstig en gunstig en kwaliteitsniveau 2 op de streefwaarde.  De betekenissen van de 2 kwaliteitsniveaus voor de verschillende versies is weergegeven in de tabel Versie in de databank en kan opgevraagd met de functie geefVersieInfo().  Geef als parameter Kwaliteitsniveau op op basis van welk kwaliteitsniveau de berekening gemaakt moet worden (strikt genomen is de berekening van de LSVI de berekening volgens kwaliteitsniveau 1).}

\item{Server}{de server waarop de databank staat die aangeroepen wordt (standaard "inbosql03\\prd")}

\item{Databank}{de naam van de databank die aangeroepen wordt (standaard "D0122_00_LSVIHabitatTypes")}

\item{Gebruiker}{gebruiker van de databank, standaard een gebruiker met leesrechten, andere opties zijn "pc-eigenaar" waarbij de login en wachtwoord gebruikt wordt waarmee ingelogd is op de pc of een eigen gebruikersnaam en wachtwoord ingeven}

\item{Wachtwoord}{wachtwoord van de gebruiker van de databank, standaard een gebruiker met leesrechten, moet niet ingevuld worden als Gebruiker "pc-eigenaar" is}
}
\value{
Deze functie genereert de resultaten in de vorm van een tabel met de meetresultaten en scores per indicator.
}
\description{
Deze functie bepaalt de Lokale Staat van Instandhouding van het gevraagde habitatsubtype op basis van de opgegeven opnames.
}
\examples{
berekenLSVI(Versie = "Versie 3", Kwaliteitsniveau = "1")

}

