% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenLSVIgooglesheet.R
\name{berekenLSVIgooglesheet}
\alias{berekenLSVIgooglesheet}
\title{Berekent de LSVI op basis van de opnames in googlesheet}
\usage{
berekenLSVIgooglesheet(Versie = geefUniekeWaarden("Versie", "VersieLSVI"),
  Kwaliteitsniveau = c("alle", "1", "2"), Data_indicatoren, Data_soorten,
  Server = "inbosql03\\\\prd", Databank = "D0122_00_LSVIHabitatTypes",
  Gebruiker = "D0122_AppR",
  Wachtwoord = "***REMOVED***")
}
\arguments{
\item{Versie}{De versie van het LSVI-rapport op basis waarvan de berekening gemaakt wordt, bv. "Versie 2" of "Versie 3".  Bij de default "alle" wordt de LSVI volgens de verschillende versies berekend.}

\item{Kwaliteitsniveau}{Voor elke versie van de LSVI zijn er een of meerdere kwaliteitsniveaus gedefinieerd in de databank.  Zo is er bij Versie 2.0 een onderscheid gemaakt tussen goede staat (A), voldoende staat (B) en gedegradeerde staat (C).  Hier duidt kwaliteitsniveau 1 de grens tussen voldoende (B) en gedegradeerd (C) aan en kwaliteitsniveau 2 het onderscheid tussen goed (A) en voldoende (B).  Bij Versie 3 duidt kwaliteitsniveau 1 op het onderscheid tussen ongunstig en gunstig en kwaliteitsniveau 2 op de streefwaarde.  De betekenissen van de 2 kwaliteitsniveaus voor de verschillende versies is weergegeven in de tabel Versie in de databank en kan opgevraagd met de functie geefVersieInfo().  Geef als parameter Kwaliteitsniveau op op basis van welk kwaliteitsniveau de berekening gemaakt moet worden (strikt genomen is de berekening van de LSVI de berekening volgens kwaliteitsniveau 1).}

\item{Data_indicatoren}{Gegevens over de indicatoren in de vorm van een data.frame met velden ID, Habitatsubtype, Criterium, Indicator en Waarde.}

\item{Data_soorten}{Bedekkingen van de sleutelsoorten in de vorm van een data.frame met velden ID, Habitatsubtype, Soort_NL of Soort_Latijn en Bedekking.}

\item{Server}{de server waarop de databank staat die aangeroepen wordt (standaard "inbosql03\\prd")}

\item{Databank}{de naam van de databank die aangeroepen wordt (standaard "D0122_00_LSVIHabitatTypes")}

\item{Gebruiker}{gebruiker van de databank, standaard een gebruiker met leesrechten, andere opties zijn "pc-eigenaar" waarbij de login en wachtwoord gebruikt wordt waarmee ingelogd is op de pc of een eigen gebruikersnaam en wachtwoord ingeven}

\item{Wachtwoord}{wachtwoord van de gebruiker van de databank, standaard een gebruiker met leesrechten, moet niet ingevuld worden als Gebruiker "pc-eigenaar" is}
}
\value{
Deze functie genereert de resultaten in de vorm van een tabel met de meetresultaten en scores per indicator.
}
\description{
Deze functie bepaalt de Lokale Staat van Instandhouding voor alle gegevens in de opgegeven googlesheet-file.  Ze maakt hiervoor gebruik van de generieke berekenfunctie berekenLSVI().  Of dit was toch de bedoeling.  Doordat in de googlesheet spaties gebruikt zijn in de invoer en ID niet de eerste kolom, is het niet evident om deze rechtstreeks in te lezen in R.  We gebruiken daarom een 'opgepoetste versie' van een deel van de gegevens.  Die opgepoetste versie blijkt ook nog lastig om om te zetten, misschien beter wachten op de databank en voorlopig werken met al gelayoute gegevens.

Voor elk type gegevens (elke databank) moet een gelijkaardig script als dit geschreven worden om de gegevens naar de juiste vorm om te zetten.  Dit om te ondervangen dat voor sommige gegevens de bedekkingen per indicator ingeschat zijn, voor andere per soort.  Dit biedt ook de mogelijkheid om de bedekkingsschalen te vertalen enz.
}
\examples{
\dontrun{
Data_indicatoren <- read.csv2(system.file("vbdata/opname_4010_gelayout_indicatoren.csv", package = "LSVI"), stringsAsFactors = FALSE)
Data_soorten <- read.csv2(system.file("vbdata/opname_4010_gelayout_soorten.csv", package = "LSVI"), stringsAsFactors = FALSE)
berekenLSVIgooglesheet(Versie = "Versie 3", Kwaliteitsniveau = "1", Data_indicatoren, Data_soorten)
}

}

