% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenStatus.R
\name{berekenStatus}
\alias{berekenStatus}
\title{Berekent de Status voor de records van een een opgegeven tabel}
\usage{
berekenStatus(Statustabel)
}
\arguments{
\item{Statustabel}{Dataframe met velden \code{Rijnr}, \code{RefMin}, \code{RefMax},
\code{Operator}, \code{WaardeMin} en \code{WaardeMax}.}
}
\value{
Deze functie geeft een tabel terug met velden \code{Rijnr} en \code{Status}
}
\description{
Deze functie, die bedoeld is als hulpfunctie voor de
hoofdfunctie \code{berekenLSVIbasis()}, evalueert de status van de records van een
opgegeven \code{Statustabel} met velden \code{Waarde}, \code{Referentiewaarde} en
\code{Operator}.
}
