% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenVerschilscores.R
\name{berekenVerschilscores}
\alias{berekenVerschilscores}
\title{Berekent de verschilscores voor de records van een een opgegeven tabel}
\usage{
berekenVerschilscores(Statustabel)
}
\arguments{
\item{Statustabel}{Dataframe met velden \code{Rijnr}, \code{RefMin}, \code{RefMax},
\code{Operator}, \code{WaardeMin}, \code{WaardeMax}, \code{TheoretischMaximum} en
\code{TypeVariabele}.}
}
\value{
Deze functie geeft een tabel terug met velden \code{Rijnr} en
\code{Verschilscore}
}
\description{
Deze functie, die bedoeld is als hulpfunctie voor de
hoofdfunctie \code{berekenLSVIbasis()}, berekent de verschilscores van de records
van een opgegeven \code{Statustabel} met velden \code{Rijnr}, \code{RefMin}, \code{RefMax},
\code{Operator}, \code{WaardeMin}, \code{WaardeMax} en \code{TheoretischMaximum}.
De verschilscores hebben een
waarde tussen \code{-1} en \code{+1} en geven negatieve of positieve afwijking ten
opzichte van de referentiewaarde.
}
