% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deselecteerKenmerkenInOpname.R
\name{deselecteerKenmerkenInOpname}
\alias{deselecteerKenmerkenInOpname}
\title{Controle van de ingevoerde opname}
\usage{
deselecteerKenmerkenInOpname(
  Kenmerken,
  Soortengroep,
  Studiegroep,
  SubAnalyseVariabele,
  SubRefMin,
  SubRefMax,
  SubOperator
)
}
\arguments{
\item{Kenmerken}{dataframe met alle opgegeven kenmerken, met velden
\code{Kenmerk}, \code{TypeKenmerk}, \code{WaardeMin} en \code{WaardeMax}}

\item{Soortengroep}{dataframe met de soortenlijst die uit Kenmerken
gedeselecteerd moet worden}

\item{Studiegroep}{dataframe met de lijst kenmerken die uit Kenmerken
gedeselecteerd moet worden.
Als ook Soortengroep opgegeven is, geeft Studiegroep aan welke kenmerken
wel behouden moeten blijven na deselecteren van de soortengroep.}

\item{SubAnalyseVariabele}{heeft waarde "bedekking" als er een subvoorwaarde
is voor de bedekking van de geselecteerde soorten of kenmerken}

\item{SubRefMin}{minimumwaarde van de grenswaarde voor de bedekking}

\item{SubRefMax}{maximumwaarde van de grenswaarde voor de bedekking}

\item{SubOperator}{operator voor deze subvoorwaarde: moet de bedekking hoger
of lager liggen dan de opgegeven referentiewaarde?}
}
\value{
Deze functie geeft een aangepaste tabel Data_soorten terug waarin
enkel de soorten uit de soortenlijst(en) opgenomen zijn en die bovendien
gekoppeld is aan de gegevens van de soortenlijst.
}
\description{
Deze hulpfunctie voor de s4-klassen \code{aantal} en \code{bedekking}
selecteert soorten of kenmerken uit een opname die niet tot de soortgroep of
studiegroep van een bepaalde voorwaarde behoren.  Op basis hiervan kunnen de
s4-klassen \code{bedekkingExcl} en \code{maxBedekkingExcl} berekend worden
(bv. dominantie van een soort:
maximale bedekking van soorten in een opname exclusief de sleutelsoorten).
}
