% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefInvoervereisten.R
\name{geefInvoervereisten}
\alias{geefInvoervereisten}
\title{Geeft de invoervereisten voor de waarde van een opname}
\usage{
geefInvoervereisten(
  Versie = "alle",
  Habitatgroep = "alle",
  Habitattype = "alle",
  Criterium = "alle",
  Indicator = "alle",
  Kwaliteitsniveau = "alle",
  Weergave = c("basis", "uitgebreid"),
  ConnectieLSVIhabitats = NULL
)
}
\arguments{
\item{Versie}{De versie van het LSVI-rapport, bv. "Versie 2" of "Versie 3".
Bij de default "alle" worden de gegevens voor de verschillende versies
gegeven.  De mogelijke waarden kunnen opgevraagd worden via
\code{geefUniekeWaarden("Versie", "VersieLSVI")} of \code{geefVersieInfo()}.}

\item{Habitatgroep}{Parameter waarmee alle habitats van een bepaalde
habitatgroep kunnen geselecteerd worden, bv. "Bossen", "Heiden",
"(Half-)natuurlijke graslanden", "Zoete en brakke wateren",...   en "alle"
(=default).  Deze waarde moet niet gespecifieerd worden als een bepaald
habitat(sub)type geselecteerd wordt.  De mogelijke waarden kunnen opgevraagd
worden via \code{geefUniekeWaarden("Habitatgroep", "Naam")}.}

\item{Habitattype}{Parameter waarmee een habitattype of habitatsubtype kan
geselecteerd worden.  Als dit een habitattype betreft met meerdere subtypes,
zullen de gegevens van alle subtypes van dit habitattype weergegeven worden.
De mogelijke waarden kunnen opgevraagd worden via
\code{geefUniekeWaarden("Habitattype", "Code")}.  Er is voor deze parameter ook de
mogelijkheid om een vector van meerdere habitat(sub)typen op te geven.}

\item{Criterium}{Het LSVI-criterium waarvoor de gegevens geselecteerd
worden: "Vegetatie", "Structuur", "Verstoring" of "alle".}

\item{Indicator}{De indicator waarvoor de gegevens uit de databank gehaald
worden.  De mogelijke waarden kunnen opgevraagd worden via
\code{geefUniekeWaarden("Indicator", "Naam")}.}

\item{Kwaliteitsniveau}{Voor elke versie van de LSVI zijn er een of meerdere
kwaliteitsniveaus gedefinieerd in de databank.  Zo is er bij Versie 2.0 een
onderscheid gemaakt tussen goede staat (A), voldoende staat (B) en
gedegradeerde staat (C).  Hier duidt kwaliteitsniveau 1 de grens tussen
voldoende (B) en gedegradeerd (C) aan en kwaliteitsniveau 2 het onderscheid
tussen goed (A) en voldoende (B).  Bij Versie 3 duidt kwaliteitsniveau 1 op
het onderscheid tussen ongunstig en gunstig en kwaliteitsniveau 2 op de
streefwaarde (uiteindelijk niet opgenomen in rapport).  De betekenissen van
de 2 kwaliteitsniveaus voor de verschillende versies is weergegeven in de
tabel Versie in de databank en kan opgevraagd met de functie
\code{geefVersieInfo()}.  Geef als parameter Kwaliteitsniveau op op basis van welk
kwaliteitsniveau de berekening gemaakt moet worden.  (Strikt genomen is de
berekening van de LSVI de berekening volgens kwaliteitsniveau 1.)}

\item{Weergave}{Wat moet er in de tabel weergegeven worden?  De default
"basis" geeft een meer overzichtelijke tabel waarbij mogelijke invoerwaarden
gescheiden door een komma in 1 cel weergegeven worden, "uitgebreid" geeft
deze invoerwaarden met alle bijhorende informatie weer in aparte records,
waardoor de tabel groot en onoverzichtelijk is.}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor
de LSVI van habitats, in te stellen d.m.v. functie \code{connecteerMetLSVIdb()}.}
}
\value{
Deze functie geeft een tabel met de hierboven beschreven informatie
uit de databank.
}
\description{
Deze functie geeft alle informatie die nodig is om
veldobservaties klaar te maken voor de berekening van de de Lokale Staat van
Instandhouding met de functie \code{berekenLSVIbasis()}, alsook de
berekeningsregels die gebruikt worden.
Allereerst geeft ze de \code{Voorwaarde} die vermeld moet worden bij de
observaties (zie \code{Data_voorwaarden} bij \code{berekenLSVIbasis()}), samen
met informatie uit de LSVI-tabellen (vnl. beoordelingsmatrix) en een
beschrijving van de voorwaarde (\code{Voorwaarde}) die zou moeten toelaten om de
koppeling te maken.

Verder geeft ze informatie over de \code{Waarde} die verwacht wordt in de functie
\code{berekenLSVIbasis()}.  \code{AnalyseVariabele} is een korte omschrijving voor de
variabele waarde, bv. "aantal" staat voor het aantal soorten of klassen en
"bedekking" voor de totale bedekking van de lijst soorten of klassen.
(Zie \code{vignette("berekeningen", package = "LSVI")} voor een uitgebreide
beschrijving van de achterliggende berekeningen.)
\code{Referentiewaarde} en \code{Operator} geven respectievelijk de grenswaarde en de
vergelijking aan op basis waarvan de beoordeling van de waarde zal gebeuren.
Voor elke \code{AnalyseVariabele} wordt informatie gegeven over het formaat dat
verwacht wordt voor Waarde: de \code{Eenheid} (die niet opgenomen moet worden in
Waarde maar wel de grootte-orde van het verwachte getal aangeeft), het
formaat van de variabele (\code{TypeVariabele}), en bij categorische variabelen
het \code{Invoertype} en de \code{Invoerwaarde} (een naam voor de categorische
variabele en de mogelijke waarden die deze kan aannemen).

Waar nodig, wordt een soortengroep of studiegroep opgegeven.  Een
studiegroep is eigenlijk equivalent aan een soortengroep: de verschillende
klassen of fasen of ... waarvoor een bedekking of andere analysevariabele
moet berekend worden.  Voorbeelden zijn groeiklassen, vegetatielagen, ...
Omwille van de overzichtelijkheid van de tabel is voor de Soortengroep enkel
een ID gegeven, de volledige lijst kan opgevraagd worden met de functie
\code{geefSoortenlijstInvoerniveau()}.

Ingeval van de \code{AnalyseVariabele} \code{aantal} kan er ook een
\code{SubAnalyseVariabele}
vermeld zijn, meestal \code{bedekking}, die aangeeft aan welke voorwaarde elke
soort of klasse afzonderlijk moet voldoen.  Aan deze \code{SubAnalysevariabele}
zijn dezelfde velden gekoppeld als aan \code{AnalyseVariabele}, nl.
\code{SubReferentiewaarde}, \code{SubOperator}, \code{SubEenheid}, \code{TypeSubVariabele},
\code{SubInvoertype} en \code{SubInvoerwaarde}.
Bijvoorbeeld, bij de voorwaarde "minimum 5 soorten minimum talrijk aanwezig"
zal de \code{AnalyseVariabele} \code{aantal} zijn, de \code{Referentiewaarde} "5",
de \code{Operator} ">=", \code{TypeVariabele} "Geheel getal", \code{SubAnalysevariabele}
"bedekking", \code{SubReferentiewaarde} "T", \code{SubOperator} ">=",
\code{TypeSubVariabele} "Categorie" en \code{SubInvoertype} "Beheermonitoringsschaal
2017".
}
\examples{
# Omwille van de iets langere lange duurtijd van de commando's staat bij
# onderstaande voorbeelden de vermelding 'dontrun' (om problemen te vermijden
# bij het testen van het package). Maar de voorbeelden werken en kunnen zeker
# uitgetest worden.
\dontrun{
library(LSVI)
maakConnectiePool()
geefInvoervereisten(
  Versie = "Versie 2.0",
  Habitattype = "4030",
  Kwaliteitsniveau = "1"
)
library(pool)
poolClose(ConnectiePool)
}

}
