% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefSoortenlijstLSVIfiche.R
\name{geefSoortenlijstLSVIfiche}
\alias{geefSoortenlijstLSVIfiche}
\title{Genereert soortenlijst(en) LSVI op basis van de opgegeven parameters}
\usage{
geefSoortenlijstLSVIfiche(Versie = geefUniekeWaarden("Versie", "VersieLSVI"),
  Habitatgroep = geefUniekeWaarden("Habitatgroep", "Habitatgroepnaam"),
  Habitattype = geefUniekeWaarden("Habitattype", "Habitatcode"),
  Habitatsubtype = geefUniekeWaarden("Habitatsubtype", "Habitatcode_subtype"),
  Criterium = geefUniekeWaarden("Criterium", "Naam"),
  Indicator = geefUniekeWaarden("Indicator", "Naam"))
}
\arguments{
\item{Versie}{De versie van het LSVI-rapport, bv. "Versie 2" of "Versie 3".  Bij de default "alle" worden de soortenlijsten voor de verschillende versies gegeven.}

\item{Habitatgroep}{Parameter waarmee alle habitats van een bepaalde habitatgroep kunnen geselecteerd worden, bv. "Bossen", "Heiden", "(Half-)natuurlijke graslanden", "Zoete wateren",...   en "alle" (=default).  Deze waarde moet niet gespecifieerd worden als een bepaald habitat(sub)type geselecteerd wordt.}

\item{Habitattype}{Parameter waarmee een habitattype kan geselecteerd worden.  Als dit habitattype meerdere subtypes heeft, zullen de soortenlijsten van alle subtypes van dit habitattype weergegeven worden.}

\item{Habitatsubtype}{Parameter waarmee een habitatsubtype geselecteerd kan worden.  Als deze waarde ingevuld is, is het niet nodig om de parameters Habitatgroep en Habitattype te specifieren.}

\item{Criterium}{Het LSVI-criterium waarvoor de soortenlijst gegeven wordt: "Vegetatie", "Structuur", "Verstoring" of "alle".}

\item{Indicator}{De indicator waarvoor de soortenlijst gegeven wordt.}
}
\value{
Deze functie geeft een tabel met velden Versie, Habitattype, Habitatsubtype, Criterium, Indicator, evt. Beschrijving, WetNaam, WetNaamKort en NedNaam (waarbij Beschrijving een omschrijving is voor een groep van soorten binnen eenzelfde indicator).  WetNaam is de volledige Latijnse naam inclusief auteursnaam, WetNaamKort bevat enkel genusnaam en soortnaam (zonder auteursnaam).
}
\description{
Deze functie genereert soortenlijsten (met wetenschappelijke en Nederlandse namen) die gebruikt worden voor de bepaling van de Lokale Staat van Instandhouding van de opgegeven parameters, zoals ze vermeld zijn in de habitatfiches (dus soms tot op soortniveau, soms tot op genusniveau of hoger).  In feite genereert ze een tabel met velden Versie, Habitattype, Habitatsubtype, Criterium, Indicator, evt. Beschrijving, WetNaam, WetNaamKort en NedNaam waarin de gespecificeerde parameters uitgeselecteerd zijn en waar voor andere parameters alle waarden uit de databank weergegeven zijn.  

De parameters kunnen enkel de hieronder gespecifeerde waarden bevatten en moeten als string opgegeven worden.  Default is telkens 'alle', waarbij de soortenlijsten voor alle mogelijke waarden van die parameter weergegeven worden (m.a.w. er is geen selectie voor deze parameter).
}
\examples{
geefSoortenlijstLSVIfiche(Habitattype = "4010")

}

