% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefSoortenlijstSoortniveau.R
\name{geefSoortenlijstSoortniveau}
\alias{geefSoortenlijstSoortniveau}
\title{Genereert soortenlijst(en) LSVI op basis van SoortengroepID}
\usage{
geefSoortenlijstSoortniveau(ConnectieLSVIhabitats, Soortengroeplijst,
  Soortenlijsttype = c("alle", "Soortniveau"))
}
\arguments{
\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor de LSVI van habitats, in te stellen d.m.v. functie connecteerMetLSVIdb.}

\item{Soortengroeplijst}{string waarin de SoortengroepID's na elkaar weergegeven worden, gescheiden door een komma}

\item{Soortenlijsttype}{"Soortniveau" betekent dat alle soorten worden weergegeven van de opgegeven soortgroepen, "alle" betekent dat alle soorten en alle taxonomische en morfologische groepen worden weergegeven die volledig in de opgegeven soortgroepen vallen (die aan de parameters voldoen (dus hier wordt de lijst uitgebreid met hogere taxonomische en morfologische groepen).}
}
\value{
Deze functie geeft een tabel met velden SoortengroepID, evt. Beschrijving, WetNaam, WetNaamKort en NedNaam (waarbij Beschrijving een omschrijving is voor een groep van soorten binnen eenzelfde indicator).  WetNaam is de volledige Latijnse naam inclusief auteursnaam, WetNaamKort bevat enkel genusnaam en soortnaam (zonder auteursnaam).
}
\description{
Deze functie genereert soortenlijsten (met wetenschappelijke en Nederlandse namen) uit de databank met de criteria en indicatoren voor de bepaling van de Lokale Staat van Instandhouding.  Het is in feite een hulpfunctie die voor verschillende andere functies gebruikt wordt en die de complexe zoekfunctie in de tabellen met soorten uitvoert op basis van een opgegeven SoortengroepID (en in die zin iets minder gebruiksvriendelijk is).  Voor een selectie van soortenlijsten op basis van specifieke parameters is de functie geefSoortenlijst() een beter alternatief.

Deze functie geeft voor de gespecifieerde soortengroepen per soortengroep een lijst van alle soorten die hieronder vallen.  Dit betekent dat voor de in de LSVI vermelde genera en soortengroepen als helofieten alle soorten vermeld worden die tot deze groepen behoren.  Voor een lijst van soort(groep)en op het niveau vermeld in de LSVI, wordt beter de functie geefSoortenlijstInvoerniveau() gebruikt.
}
\examples{
ConnectieLSVIhabitats <- connecteerMetLSVIdb()
geefSoortenlijstSoortniveau(ConnectieLSVIhabitats,"139,142,370,371")
geefSoortenlijstSoortniveau(ConnectieLSVIhabitats,"139,142,370,371","Soortniveau")
library(RODBC)
odbcClose(ConnectieLSVIhabitats)

}
