% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoercontroleData_voorwaarden.R
\name{invoercontroleData_voorwaarden}
\alias{invoercontroleData_voorwaarden}
\title{Invoercontrole voor dataframe Data_voorwaarden}
\usage{
invoercontroleData_voorwaarden(Data_voorwaarden, ConnectieLSVIhabitats, LIJST)
}
\arguments{
\item{Data_voorwaarden}{dataframe waarop invoercontrole moet gebeuren.}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor
de LSVI van habitats, in te stellen d.m.v. functie \code{connecteerMetLSVIdb()}.}

\item{LIJST}{Dataframe met lijst die weergeeft hoe de vertaling moet
gebeuren van categorische variabelen naar numerieke waarden (en omgekeerd).
Default worden deze waarden uit de databank met LSVI-indicatoren gehaald
d.m.v. de functie \code{vertaalInvoerInterval()}.
Aangeraden wordt om deze default
te gebruiken (dus parameter niet expliciet invullen), of deze waar nodig aan
te vullen met eigen schalen.  Omdat er ook een omzetting moet gebeuren voor
grenswaarden uit de databank, kan het niet doorgeven van een gedeelte van
deze lijst problemen geven.}
}
\description{
Om te vermijden dat we meermaals dezelfde invoercontrole moeten uitvoeren en
om de hoofdscripts overzichtelijk te houden, maken we voor elke
invoercontrole een aparte hulpfunctie aan, die we kunnen aanroepen.  Deze
wordt NIET geëxporteerd, dus deze functies kunnen niet als commando gerund
worden (maar worden wel gerund als de functie waarin ze voorkomen,
aangeroepen wordt).  Ingeval van Data_voorwaarden is ook de omzetting van de
voorwaarden naar een interval opgenomen in de functie.
}
