% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDatabankfouten.R
\name{logDatabankfouten}
\alias{logDatabankfouten}
\title{Lijst alle nog op te lossen databankfouten op}
\usage{
logDatabankfouten(ConnectieLSVIhabitats = NULL)
}
\arguments{
\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor
de LSVI van habitats, in te stellen d.m.v. functie \code{connecteerMetLSVIdb()}.}
}
\value{
Deze functie geeft een list met 2 dataframes terug
}
\description{
Deze functie maakt een tabel met alle problemen die nog in de databank
zitten.  Enerzijds is er een beperkte tabel met problemen die op een hoger
niveau opgelost kunnen worden en anderzijds een detail met alle Voorwaarden
waar nog een fout in zit.  Problemen die op beide niveaus kunnen opgelost
worden (bv. benoemen van analysevariabelen), staan op beide niveaus vermeld.
}
