% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maakConnectiePool.R
\name{maakConnectiePool}
\alias{maakConnectiePool}
\title{Maak een connectiepool met de databank met LSVI-indicatoren in het
package}
\usage{
maakConnectiePool()
}
\value{
Deze functie maakt een Environment-object aan dat de connecties
regelt met de betreffende databank.
}
\description{
Deze functie maakt een connectiepool met de in het package
toegevoegde databank met LSVI-indicatoren, wat nodig is om de functies te
kunnen gebruiken.  Deze connectiepool moet eenmalig aangemaakt worden, en
functies zullen automatisch deze connectiepool gebruiken om te connecteren
met de databank.  Alternatief is om een connectie aan te maken met de
functie \code{connecteerMetLSVIdb()} en deze bij elke functie mee te geven.
}
\examples{
library(LSVI)
maakConnectiePool()
geefVersieInfo()
library(pool)
poolClose(ConnectiePool)

}
