% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maakConnectiepoolSQLite.R
\name{maakConnectiepoolSQLite}
\alias{maakConnectiepoolSQLite}
\title{Maak een connectiepool de databank met LSVI-indicatoren in het package}
\usage{
maakConnectiepoolSQLite()
}
\value{
Deze functie maakt een Environment-object aan dat de connecties regelt met de betreffende databank in de installatie-file van het package.
}
\description{
\examples{
library(LSVI)
maakConnectiepoolSQLite()
geefVersieInfo()
sluitConnectiePool()

Deze functie maakt een connectiepool met de in het package toegevoegd databank met LSVI-indicatoren, wat nodig is om de functies te kunnen gebruiken.  Deze connectiepool moet eenmalig aangemaakt worden, en functies zullen automatisch deze connectiepool gebruiken om te connecteren met de databank.  Alternatief is om een connectie aan te maken met de functie connecteerMetLSVIlite() en deze bij elke functie mee te geven.  Voorlopig bevat deze databank bij wijze van test maar enkele gegevens en wordt best connectie gemaakt met de volledige databank via de functie maakConnectiePool().
}
