% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_maxBedekking.R
\docType{class}
\name{maxBedekking-class}
\alias{maxBedekking-class}
\title{S4-klasse die de maximale bedekking van de soorten binnen een soortengroep
berekent}
\description{
Deze klasse MaxBedekking staat in voor de berekening van waarden voor
TypeVariabele Bedekking op basis van opgegeven kenmerken.  Ze is een
nakomeling van de klasse AnalyseVariabele.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk,
TypeKenmerk, WaardeMin en WaardeMax}
}}

