% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_maxBedekkingExcl.R
\docType{class}
\name{maxBedekkingExcl-class}
\alias{maxBedekkingExcl-class}
\title{S4-klasse die de maximale bedekking berekent van de soorten die niet tot de
soortengroep behoren}
\description{
Deze klasse MaxBedekkingExcl staat in voor de berekening van waarden voor
TypeVariabele Bedekking op basis van opgegeven kenmerken.  Ze is een
nakomeling van de klasse AnalyseVariabele.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk,
TypeKenmerk, WaardeMin en WaardeMax}
}}

