% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_meting.R
\docType{class}
\name{meting-class}
\alias{meting-class}
\title{S4-klasse voor voorwaarden die enkel rechtstreeks op het terrein gemeten kunnen worden of afgeleid worden uit GIS-data}
\description{
Deze klasse meting geeft een waarde NA.  Ze is een nakomeling van de klasse AnalyseVariabele.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

