% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTaxonnaam.R
\name{parseTaxonnaam}
\alias{parseTaxonnaam}
\title{Vereenvoudig de soortnaam door de auteurs te verwijderen}
\usage{
parseTaxonnaam(Taxonnaam, ParseType = "canonicalnamewithmarker")
}
\arguments{
\item{Taxonnaam}{Wetenschappelijke naam of namen die vereenvoudigd moeten
worden (String)}

\item{ParseType}{Welk type parsing teruggegeven moet worden.  Standaard is
dit \code{"canonicalnamewithmarker"}, andere opties zijn \code{"canonicalname"} en
\code{"canonicalnamecomplete"}}
}
\value{
Deze functie geeft de vereenvoudigde naam of namen terug (als string
of vector van strings)
}
\description{
\strong{Deze functie wordt niet meer onderhouden en zal bij een volgende versie
van dit package verdwijnen.
Voor een functie met gelijkaardige functionaliteit verwijzen we naar
\code{rgbif::name_parse()} waarrond \code{parseTaxonnaam()} een wrapper is.}

Deze functie vereenvoudigt de opgegeven taxonnaam of taxonnamen door
de auteursnaam te verwijderen.  Ze is gebaseerd op de functie \code{name_parse()}
uit het package rgbif, maar ze vangt enkele specifieke situaties op uit
plantenlijsten die op INBO gebruikt worden en die de originele functie niet
correct opvangt, bv.
\code{"v."}, \code{"an"} en \code{"den"} herkennen als deel van een auteursnaam en "s.l." en
meerdere soortnamen gescheiden door \code{"/"} (zonder spatie) vervangen door een
achtervoegsel "groep".
}
\examples{
parseTaxonnaam("Rosa canina L.")

}
