% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecteerSoortenInOpname.R
\name{selecteerSoortenInOpname}
\alias{selecteerSoortenInOpname}
\title{Controle van de ingevoerde opname}
\usage{
selecteerSoortenInOpname(Data_soorten, Soortengroeplijst)
}
\arguments{
\item{Data_soorten}{Bedekkingen van de sleutelsoorten in de vorm van een data.frame met velden ID en Soort_NL of Soort_Latijn.  Eventueel andere velden worden gewoon teruggegeven bij de uitvoer. (Eventueel zou hier ook de NBNTaxonVersionKey kunnen gebruikt worden.)}

\item{Soortengroeplijst}{string waarin de SoortengroepID's na elkaar weergegeven worden, gescheiden door een komma}
}
\value{
Deze functie geeft een aangepaste tabel Data_soorten terug waarin enkel de soorten uit de soortenlijst(en) opgenomen zijn en die bovendien gekoppeld is aan de gegevens van de soortenlijst.
}
\description{
Deze hulpfunctie controleert of de ingevoerde opname geen fouten bevat en koppelt ze meteen aan de opgegeven soortenlijst.  Ze controleert of veld ID en een veld met een soortenlijst (soort_Latijn of soort_NL) aanwezig is, en controleert voor de soortenlijst of de namen wel bestaan (conform zijn met de LSVI-databank).  Problemen hiermee geven een error.  Verder controleert ze of alle soorten uit de opgegeven soortengroep(en) in de opname aanwezig zijn en ze geeft een warning als dit niet zo is.
}

