% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sluitConnectiePool.R
\name{sluitConnectiePool}
\alias{sluitConnectiePool}
\title{Sluit een eerder aangemaakte connectiepool}
\usage{
sluitConnectiePool()
}
\value{
Er wordt geen waarde teruggegeven, de connectie wordt gesloten.
}
\description{
Deze functie zorgt ervoor dat een door de functie maakConnectiePool() aangemaakte connectiepool gesloten wordt, wat betekent dat de verbinding met de databank verbroken wordt en de functies van het LSVI-package niet meer gerund kunnen worden.  Het Environment-object ConnectiePool blijft wel bestaan, maar is dus inactief.  Deze functie is handig als er in een script slechts kort gebruik gemaakt wordt van het LSVI-package.
}
\examples{
library(LSVI)
maakConnectiePool()
geefVersieInfo()
sluitConnectiePool()

}
