% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertaalBedekkingTansley.R
\name{vertaalBedekkingTansley}
\alias{vertaalBedekkingTansley}
\title{Vertaalt bedekking in procent naar Tansley-waarde}
\usage{
vertaalBedekkingTansley(Percentages,
  Tansley = read_csv(system.file("schaaltabellen/Tansley.csv", package =
  "LSVI")))
}
\arguments{
\item{Percentages}{vector met percentages, dus getallen tussen 0 en 100, die omgezet moeten worden naar Tansley-waarden}

\item{Tansley}{tabel met kolommen 'Voluit' (= namen Tansley-schaal), 'Ondergrens' en 'Bovengrens'.  Default wordt de in het package meegeleverde tabel gebruikt (Tansley.csv in de map schaaltabellen).}
}
\value{
Vector met de Tansley-waarden van de ingevoerde percentages
}
\description{
Deze functie vertaalt een ingevoerde vector van getallen (bedekkingen in procent) naar de categorieen uit de Tansley-schaal op basis van een tabel waarin onder- en bovengrenzen voor de schaalverdeling gegeven zijn.
}
\examples{
\dontrun{
vertaalBedekkingTansley(c(2.5, 10, 30))
}


}
