% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertaalIntervalUitvoer.R
\name{vertaalIntervalUitvoer}
\alias{vertaalIntervalUitvoer}
\title{zet een interval om naar een waarde in de gevraagde eenheid}
\usage{
vertaalIntervalUitvoer(Dataset, LIJST, ConnectieLSVIhabitats)
}
\arguments{
\item{Dataset}{dataframe met velden \code{Rijnr}, \code{Type}, \code{Min}, \code{Max}, \code{Eenheid}
en \code{Invoertype}}

\item{LIJST}{Dataframe met lijst die weergeeft hoe de vertaling moet
gebeuren van numerieke waarden naar categorische variabelen.  Verschillend
van andere functies die dezelfde lijst gebruiken, mogen hier geen
overlappende categorieën voorkomen binnen eenzelfde schaal.  Om zulke lijst
te bekomen, moeten uit de lijst gegenereerd door de functie
\code{vertaalInvoerInterval()} de records met Basisschaal 1 gefilterd worden.}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor
de LSVI van habitats, in te stellen d.m.v. functie \code{connecteerMetLSVIdb()}.}
}
\value{
Dataframe met velden Min
}
\description{
Deze functie zet een interval bestaande uit minimumwaarde en
maximumwaarde om naar een uitvoerwaarde in de opgegeven eenheid.
De functie gebruikt \code{Type}, \code{Eenheid} en \code{Invoertype} om te bepalen welke
omzetting eventueel nodig is.
Als minimum en maximum niet dezelfde waarde hebben, geeft ze
beide waarden weer, gescheiden door een "-".
}
