% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertaalTansleyPercentage.R
\name{vertaalTansleyPercentage}
\alias{vertaalTansleyPercentage}
\title{Vertaalt Tansley-waarde naar bedekking in procent}
\usage{
vertaalTansleyPercentage(TansleyVector,
  Tansley = read_csv(system.file("schaaltabellen/Tansley.csv", package =
  "LSVI")))
}
\arguments{
\item{TansleyVector}{vector met percentages, dus getallen tussen 0 en 100, die omgezet moeten worden naar Tansley-waarden}

\item{Tansley}{tabel met kolommen 'Voluit' (= namen Tansley-schaal) en 'GemBedekking' (= klassemiddens van de bedekkingsschaal).  Default wordt de in het package meegeleverde tabel gebruikt (Tansley.csv in de map schaaltabellen).}
}
\value{
Vector met een overeenkomstig bedekkingspercentage voor elke categorie vermeld in Tansleyvector
}
\description{
Deze functie vertaalt een ingevoerde vector van categorieen uit de Tansley-schaal naar bedekkingen in procent op basis van een tabel waarin de gemiddelde bedekkingen voor de schaalverdelingen gegeven zijn.
}
\examples{
vertaalTansleyPercentage(c("sporadisch", "frequent", "codominant"))

}

