% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_aandeelKruidlaag.R,
%   R/s4_aandeelVegetatielaag.R
\docType{class}
\name{aandeelKruidlaag-class}
\alias{aandeelKruidlaag-class}
\title{S4-klasse die aandeelVegetatielaag bedekking soorten binnen een vegetatielaag berekend}
\description{
Deze klasse AandeelVegetatielaag staat in voor de berekening van waarden voor TypeVariabele AandeelVegetatielaag op basis van opgegeven kenmerken.  Ze is een nakomeling van de klasse AnalyseVariabele.

Deze klasse AandeelVegetatielaag staat in voor de berekening van waarden voor TypeVariabele AandeelVegetatielaag op basis van opgegeven kenmerken.  Ze is een nakomeling van de klasse AnalyseVariabele.
}
\section{Slots}{

\describe{
\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}

\item{\code{Kenmerken}}{dataframe met alle opgegeven kenmerken, met velden Kenmerk, TypeKenmerk, WaardeMin en WaardeMax}
}}

