% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analyseVariabele_c.R
\name{analyseVariabele_c}
\alias{analyseVariabele_c}
\title{Constructor van s4-classe AnalyseVariabele}
\usage{
analyseVariabele_c(VoorwaardeID, Kenmerken, ConnectieLSVIhabitats, ConnectieNBN,
  LIJST)
}
\description{
Constructor van s4-classe AnalyseVariabele
}
