% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analyseVariabele_c.R
\name{analyseVariabele_c}
\alias{analyseVariabele_c}
\title{Constructor van s4-classe AnalyseVariabele}
\usage{
analyseVariabele_c(VoorwaardeID, Kenmerken, ConnectieLSVIhabitats, ConnectieNBN,
  LIJST)
}
\description{
Een constructor is een gebruiksvriendelijke functie die een s4-klasse aanmaakt, zodat een gebruiker niet rechtstreeks geconfronteerd wordt met het aanmaken van een object voor een s4-klasse.  In dit geval worden alle als parameter toegevoegde gegevens netjes in het object gestoken, alsook extra info die uit de databank gehaald wordt.  Een deel van de validatie gebeurt in de s4-klasse AnalyseVariabele zelf.
}
