% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenLSVIbasis.R
\name{berekenLSVIbasis}
\alias{berekenLSVIbasis}
\title{Berekent de LSVI op basis van VoorwaardeID en opgegeven waarden}
\usage{
berekenLSVIbasis(
  Versie = "alle",
  Kwaliteitsniveau = "alle",
  Data_habitat,
  Data_voorwaarden = data.frame(ID = character(), Criterium = character(), Indicator =
    character(), Voorwaarde = character(), Waarde = character(), Type = character(),
    WaardeMin = double(), WaardeMax = double(), stringsAsFactors = FALSE),
  Data_soortenKenmerken = data.frame(ID = character()),
  Aggregatiemethode = "1-out-all-out",
  ConnectieLSVIhabitats = NULL,
  LIJST = geefVertaallijst(ConnectieLSVIhabitats),
  na.rm = FALSE
)
}
\arguments{
\item{Versie}{De versie van het LSVI-rapport op basis waarvan de berekening
gemaakt wordt, bv. "Versie 2.0" of "Versie 3".  Bij de default "alle" wordt
de LSVI volgens de verschillende versies berekend.}

\item{Kwaliteitsniveau}{Voor elke versie van de LSVI zijn er een of meerdere
kwaliteitsniveaus gedefinieerd in de databank.  Zo is er bij Versie 2.0 een
onderscheid gemaakt tussen goede staat (A), voldoende staat (B) en
gedegradeerde staat (C).  Hier duidt kwaliteitsniveau 1 de grens tussen
voldoende (B) en gedegradeerd (C) aan en kwaliteitsniveau 2 het onderscheid
tussen goed (A) en voldoende (B).  Bij Versie 3 duidt kwaliteitsniveau 1 op
het onderscheid tussen ongunstig en gunstig en kwaliteitsniveau 2 op de
streefwaarde (uiteindelijk niet opgenomen in rapport).  De betekenissen van
de 2 kwaliteitsniveaus voor de verschillende versies is weergegeven in de
tabel Versie in de databank en kan opgevraagd met de functie
\code{geefVersieInfo()}.  Geef als parameter Kwaliteitsniveau op op basis van welk
kwaliteitsniveau de berekening gemaakt moet worden.  (Strikt genomen is de
berekening van de LSVI de berekening volgens kwaliteitsniveau 1.)}

\item{Data_habitat}{Een opsomming van de te analyseren opnamen met opgave
van het aanwezige habitattype (= het habitattype volgens welke criteria de
beoordeling moet gebeuren).  Deze info moet doorgegeven worden in de vorm
van een dataframe met minimum de velden \code{ID} en \code{Habitattype}, waarbij \code{ID}
een groeperende variabele is voor een opname (plaats en tijdstip).
\code{Habitattype}
moet overeenkomen met de naamgeving in de LSVI-databank (op te zoeken door
\code{geefUniekeWaarden("Habitattype", "Code")}).  Eventuele extra velden zullen
overgenomen worden bij de uitvoer.}

\item{Data_voorwaarden}{Gegevens over de opgemeten indicatoren in de vorm
van een dataframe met velden \code{ID}, \code{Criterium}, \code{Indicator}, \code{Voorwaarde},
\code{Waarde}, \code{Type}, \code{Invoertype} en \code{Eenheid}, waarbij \code{ID} de groeperende
variabele voor een opname is die ook bij \code{Data_habitat} opgegeven is.
\code{Criterium}, \code{Indicator} en
\code{Voorwaarde} moeten overeenkomen met de waarde in de databank (op te zoeken
via de functie \code{geefInvoervereisten()}).  \code{Waarde} is de waarde die voor die
voorwaarde geobserveerd of gemeten is en Type het soort variabele (zie
\code{geefUniekeWaarden("TypeVariabele", "Naam")} voor de mogelijke waarden).
Ingeval van een categorische variabele moet bij \code{Invoertype} de naam van de
lijst opgegeven worden waaruit deze waarde komt (bv. welke schaal gebruikt
is, zie \code{geefUniekeWaarden("Lijst", "Naam")} voor alle mogelijkheden).  Als
een indicator rechtstreeks op het veld ingeschat is, kan deze ingevoerd
worden door in deze tabel de kolom voorwaarde leeg te laten (wat in R
aangeduid wordt door NA) en als waarde "TRUE" of "FALSE" in te geven.  In
dit geval moeten \code{Type}, \code{Invoertype} en \code{Eenheid} niet ingevoerd worden.}

\item{Data_soortenKenmerken}{Gegevens van soorten en kenmerken en hun
bedekking (m.a.w. enkel kenmerken waarvan een bedekking gemeten is, horen
in deze tabel).
Deze dataframe moet de velden \code{ID}, \code{Vegetatielaag}, \code{Kenmerk},
\code{TypeKenmerk}, \code{Waarde}, \code{Type}, \code{Invoertype} en \code{Eenheid} bevatten, waarbij
\code{ID} de groeperende variabele voor een opname is die ook bij \code{Data_habitat}
opgegeven is.
\code{Kenmerk} bevat een soortnaam of een naam die voorkomt in de lijst
gegenereerd door \code{geefUniekeWaarden("StudieItem", "Waarde")} en \code{TypeKenmerk}
geeft een beschrijving voor dat kenmerk: \code{"studiegroep"}, \code{"soort_Latijn"},
\code{"soort_NL"} of \code{"soort_gbif"}.
\code{Waarde} is de geobserveerde bedekking en \code{Type} het
soort variabele dat voor de bedekking gebruikt is (zie
\code{geefUniekeWaarden("TypeVariabele", "Naam")} voor de mogelijke waarden).
Ingeval van een categorische variabele moet bij \code{Invoertype} de naam van de
lijst opgegeven worden welke schaal gebruikt is
(zie \code{geefUniekeWaarden("Lijst", "Naam")} voor alle mogelijkheden).}

\item{Aggregatiemethode}{Keuze van de methode om tot één beoordeling per
criterium of per habitatvlek of meetpunt te komen. Er zijn twee opties:
(1) "RapportageHR": de beoordeling is gunstig als meer dan 50 procent van de
indicatoren gunstig zijn EN als geen enkele zeer belangrijke indicator
ongunstig is; (2) "1-out-all-out": de beoordeling is gunstig als alle
indicatoren gunstig zijn. "1-out-all-out" is default.}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor
de LSVI van habitats, in te stellen d.m.v. functie \code{connecteerMetLSVIdb()}.}

\item{LIJST}{Dataframe met lijst die weergeeft hoe de vertaling moet
gebeuren van categorische variabelen naar numerieke waarden (en omgekeerd).
Default worden deze waarden uit de databank met LSVI-indicatoren gehaald
d.m.v. de functie \code{vertaalInvoerInterval()}.
Aangeraden wordt om deze default
te gebruiken (dus parameter niet expliciet invullen), of deze waar nodig aan
te vullen met eigen schalen.  Omdat er ook een omzetting moet gebeuren voor
grenswaarden uit de databank, kan het niet doorgeven van een gedeelte van
deze lijst problemen geven.}

\item{na.rm}{Hier geef je aan hoe de berekening moet omgaan met NA-waarden.
Default is FALSE. Dit betekent dat NA-waarden niet worden verwijderd.
Hierdoor zal de indexberekening resulteren in een NA zodra één van de
indicatoren NA is. Voor de berekening van de status zal dit enkel resulteren
in een NA indien minstens één van de indicatoren NA is en minstens één
van de indicatoren status TRUE (= gunstig) heeft. Indien \code{na.rm = TRUE}
worden
eventuele NA-waarden verwijderd zodat status en de indices een resultaat
hebben. Doordat deze dan mogelijk niet op de volledige set van indicatoren
gebaseerd zijn, moet hiermee rekening gehouden worden afhankelijk van de
context waarvoor de resultaten gebruikt worden.}
}
\value{
Deze functie genereert de resultaten in de vorm van een list met 5
tabellen: een eerste met de beoordelingen per kwaliteitsniveau, een tweede
met de beoordelingen per criterium en kwaliteitsniveau, een derde met de
beoordelingen per indicator en kwaliteitsniveau, en een vierde met de
detailgegevens inclusief meetwaarden.
Een vijfde tabel \code{Soortenlijst} bevat de taxonomische info die gebruikt is
om de opgegeven taxa te koppelen aan de taxa die deel uitmaken van de
beoordeling.
}
\description{
Deze functie bepaalt de Lokale Staat van Instandhouding en
biotische indices op basis van gegevens, die in het juiste formaat moeten
aangeleverd worden.  Zie hiervoor de beschrijving bij de parameters
(\href{https://inbo.github.io/LSVI/reference/berekenLSVIbasis.html#arguments}{"Arguments" })
en de tabellen van het voorbeeld.  In principe is enkel de
parameter \code{Data_habitat} verplicht om op te geven, maar extra datasets zijn
uiteraard wel nodig om een resultaat te bekomen.  Welke datasets relevant
zijn, is afhankelijk van de opgegeven habitattypes: voor een aantal
habitattypes kan een tabel met observaties en hun bedekking of aanwezigheid
(=parameter \code{Data_soortenKenmerken}) volstaan, voor bossen zijn bv.
bijkomend gegevens nodig over dood hout.

De Lokale Staat van Instandhouding wordt weergegeven in de kolom \code{Status}
met als mogelijke waarden TRUE (= gunstig) en FALSE (= ongunstig).

De biotische indices zijn afgeleid van het verschil tussen een geobserveerde
waarde en de referentiewaarde voor elke indicator. Deze verschillen werden
herschaald tussen \code{+1} en \code{-1}, waarbij een positieve en negatieve waarde
overeenkomt met respectievelijk een gunstige en ongunstige score. Deze
verschilscores per indicator worden geaggregeerd, eerst voor de indicatoren
die tot eenzelfde criterium behoren, vervolgens worden deze geaggregeerde
scores verder geaggregeerd om tot een globale index te komen. Er worden drie
verschillende globale indices berekend waarbij de naamgeving aangeeft welk
aggregatie achtereenvolgens gebruikt werd: \code{"index_min_min"},
\code{"index_min_harm"} en \code{"index_harm_harm"}.
Een naam met \code{"min"} duidt op minimum van de scores als
aggregatie; bij \code{"harm"} werd het harmonisch gemiddelde berekend.

Bij de output zit een tabel \code{Soortenlijst} die aangeeft aan welke taxoninfo
de opgegeven taxa gekoppeld zijn.
Deze laat dus toe om te controleren of de koppeling juist gebeurd is,
en of de bovenliggende taxonniveaus kloppen.
Zoals in vignet berekeningen (\code{vignette("Berekeningen", package = "LSVI")})
in detail beschreven is,
wordt na koppeling met de taxonlijst uit het package op ongekoppelde taxa
een matching toegepast op basis van een Gbif-tool om bij kleine afwijkingen
toch de ingevoerde soorten te herkennen.
Op die manier wordt de kans op het herkennen van de soort door het package
vergroot, maar het heeft als keerzijde dat een taxon door Gbif ten onrechte
gekoppeld kan worden aan een Gbif-key van een accepted taxon (omdat Gbif
geen rekening houdt met auteursnamen, "auct.", "auct. non", "non",...).
Daarom wordt met een warning aangeraden om de \code{Soortenlijst} te controleren
als de Gbif-tool voor de koppeling van minstens één soort gebruikt is
(dit is aangegeven in veld \code{Koppelmethode}, en deze records staan bovenaan).
Fouten bij gebruik van de Gbif-tool kunnen best opgelost worden door de taxa
bij invoer te vervangen door de schrijfwijze in de databank in het package
(beschikbaar
op \href{https://zenodo.org/records/10561497}{Zenodo}).
Fouten in deze taxonlijst mogen gemeld worden via
\href{https://github.com/inbo/LSVI/issues}{een issue} (of aan de beheerder van
het package).
}
\examples{
# Omwille van de iets langere lange duurtijd van de commando's staat bij
# onderstaand voorbeeld de vermelding 'dontrun' (om problemen te vermijden
# bij het testen van het package). Maar het voorbeeld werkt en kan zeker
# uitgetest worden.
\dontrun{
library(LSVI)
maakConnectiePool()
library(readr)
Data_habitat <-
  read_csv2(system.file("vbdata/Opname4030habitat.csv", package = "LSVI"),
    col_types = list(col_character(), col_character(),col_character()))
Data_voorwaarden <-
  read_csv2(
    system.file("vbdata/Opname4030voorwaardenv2.csv", package = "LSVI"))
Data_soortenKenmerken <-
  read_csv2(
    system.file("vbdata/Opname4030soortenKenmerken.csv", package = "LSVI"))
berekenLSVIbasis(Versie = "Versie 2.0",
                 Kwaliteitsniveau = "1", Data_habitat,
                 Data_voorwaarden, Data_soortenKenmerken)

# Hoe testen welke soorten aan welke indicatoren gekoppeld worden?
library(tidyverse)
berekenLSVIbasis(
  Versie = "Versie 2.0",
  Kwaliteitsniveau = "1", Data_habitat,
  Data_voorwaarden, Data_soortenKenmerken
 )[["Soortenlijst"]] \%>\%
   pivot_longer(
     ends_with("Key"), names_to = "Niveau", values_to = "GbifKey"
   ) \%>\%
   mutate(
     Niveau = toupper(gsub("Key", "", .data$Niveau))
   ) \%>\%
   filter(!is.na(.data$GbifKey)) \%>\%
   inner_join(
     geefSoortenlijst(
       Versie = "Versie 2.0", Habitattype = "4030",
       Taxonlijstniveau = "indicator"),
     by = c("Niveau" = "Rank", "GbifKey" = "GbifUsageKey")
   )
}

}
