% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenLSVIbasis.R
\name{berekenLSVIbasis}
\alias{berekenLSVIbasis}
\title{Berekent de LSVI op basis van VoorwaardeID en opgegeven waarden}
\usage{
berekenLSVIbasis(Versie = geefUniekeWaarden("Versie", "VersieLSVI"),
  Kwaliteitsniveau = c("alle", "1", "2"), Data_voorwaarden)
}
\arguments{
\item{Versie}{De versie van het LSVI-rapport op basis waarvan de berekening gemaakt wordt, bv. "Versie 2" of "Versie 3".  Bij de default "alle" wordt de LSVI volgens de verschillende versies berekend.}

\item{Kwaliteitsniveau}{Voor elke versie van de LSVI zijn er een of meerdere kwaliteitsniveaus gedefinieerd in de databank.  Zo is er bij Versie 2.0 een onderscheid gemaakt tussen goede staat (A), voldoende staat (B) en gedegradeerde staat (C).  Hier duidt kwaliteitsniveau 1 de grens tussen voldoende (B) en gedegradeerd (C) aan en kwaliteitsniveau 2 het onderscheid tussen goed (A) en voldoende (B).  Bij Versie 3 duidt kwaliteitsniveau 1 op het onderscheid tussen ongunstig en gunstig en kwaliteitsniveau 2 op de streefwaarde.  De betekenissen van de 2 kwaliteitsniveaus voor de verschillende versies is weergegeven in de tabel Versie in de databank en kan opgevraagd met de functie geefVersieInfo().  Geef als parameter Kwaliteitsniveau op op basis van welk kwaliteitsniveau de berekening gemaakt moet worden (strikt genomen is de berekening van de LSVI de berekening volgens kwaliteitsniveau 1).}

\item{Data_voorwaarden}{Gegevens over de opgemeten indicatoren in de vorm van een data.frame met velden ID, Habitatsubtype, VoorwaardeID en Waarde, waarbij ID een groeperende variabele is voor een opname (plaats en tijdstip) en Waarde de waarde die voor die indicator geobserveerd of gemeten is.  Het type van deze waarde moet overeenkomen met het type dat verwacht wordt volgens de LSVI (geheel getal als een aantal (soorten) verwacht wordt, decimaal getal tussen 0 en 100 als een percentage verwacht wordt, een van de mogelijke categorieen bij een categorische variabele,...)  VoorwaardeID komt overeen met de ID in de databank die gekoppeld is aan de voorwaarde en Habitatsubtype moet overeenkomen met de naamgeving in de LSVI-databank (op te zoeken door geefUniekeWaarden("Habitatsubtype", "Habitatcode_subtype")).}
}
\value{
Deze functie genereert de resultaten in de vorm van een tabel met de meetresultaten en scores per indicator.
}
\description{
Deze hulpfunctie bepaalt de Lokale Staat van Instandhouding op basis van een opgegeven tabel met VoorwaardeID en een opgegeven waarde (die in het juiste formaat moet zijn).  Ze is bedoeld als technische hulpfunctie, een gelijkaardige functie met meer mogelijkheden voor invoer is berekenLSVI.
}

