% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berekenStatus.R
\name{berekenStatus}
\alias{berekenStatus}
\title{Berekent de Status voor de records van een een opgegeven tabel}
\usage{
berekenStatus(Statustabel)
}
\arguments{
\item{Statustabel}{Tabel met velden ...}
}
\value{
Deze functie voegt een extra kolom Status toe aan de opgegeven tabel en verwijdert de velden ...
}
\description{
Deze functie, die bedoeld is als hulpfunctie voor de hoofdfunctie berekenLSVIbasis, voegt een kolom Status toe aan de opgegeven tabel waarin ze de status evalueert op basis van onder andere de velden Waarde, Referentiewaarde en Operator (zie parameter Statustabel) en ze wist ook enkele velden.
}
