% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinerenVoorwaarden.R
\name{combinerenVoorwaarden}
\alias{combinerenVoorwaarden}
\title{combineert de Status van voorwaarden via de opgegeven formule}
\usage{
combinerenVoorwaarden(Formule, VoorwaardeID, Status)
}
\arguments{
\item{Formule}{string van ID's gecombineerd met EN en OF, bijvoorbeeld '(720 EN 721) OF 15'}

\item{VoorwaardeID}{vector van alle voorwaardeID's die voorkomen in de Formule}

\item{Status}{vector met voor elke VoorwaardeID een overeenkomstige logische waarde status (TRUE of FALSE)}
}
\value{
logische waarde TRUE/FALSE die de uitkomst van de Formule is (gecombineerd met VoorwaardeID en Status)
}
\description{
Nog aanvullen
}
\examples{


}
