% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecteerMetLSVIdb.R
\name{connecteerMetLSVIdb}
\alias{connecteerMetLSVIdb}
\title{Connecteer met de databank met LSVI-indicatoren in het package}
\usage{
connecteerMetLSVIdb()
}
\value{
Deze functie geeft een open odbc-connectie naar de SQLite-databank
in de installatie-file van het package.
}
\description{
Deze functie maakt een connectie met de in het package
toegevoegde databank met LSVI-indicatoren, wat nodig is om de functies te
kunnen gebruiken.  Deze connectie moet als argument meegegeven worden bij
elke functie functie die informatie uit de databank ophaalt.  Alternatief is
om eenmalig een connectiepool aan te maken met de functie
maakConnectiePool().
}
\examples{
library(LSVI)
ConnectieLSVIhabitats <- connecteerMetLSVIdb()
geefVersieInfo(ConnectieLSVIhabitats)
library(DBI)
dbGetQuery(ConnectieLSVIhabitats,
           "SELECT VersieLSVI, Referentie FROM Versie")
dbDisconnect(ConnectieLSVIhabitats)

}
