% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecteerMetLSVIdb.R
\name{connecteerMetLSVIdb}
\alias{connecteerMetLSVIdb}
\title{Connecteer met de databank met LSVI-indicatoren}
\usage{
connecteerMetLSVIdb(Server = "inbosql03\\\\prd",
  Databank = "D0122_00_LSVIHabitatTypes", Gebruiker = "D0122_AppR",
  Wachtwoord = "19D939F1-BCCE-439F-9ED4-6A886E038A6D")
}
\arguments{
\item{Server}{de server waarop de databank staat die aangeroepen wordt (standaard "inbosql03\\prd")}

\item{Databank}{de naam van de databank die aangeroepen wordt (standaard "D0122_00_LSVIHabitatTypes")}

\item{Gebruiker}{gebruiker van de databank, standaard een gebruiker met leesrechten, andere opties zijn "pc-eigenaar" waarbij de login en wachtwoord gebruikt wordt waarmee ingelogd is op de pc of een eigen gebruikersnaam en wachtwoord ingeven}

\item{Wachtwoord}{wachtwoord van de gebruiker van de databank, standaard een gebruiker met leesrechten, moet niet ingevuld worden als Gebruiker "pc-eigenaar" is}
}
\value{
Deze functie geeft een open odbc-connectie naar de gespecifieerde databank.
}
\description{
Deze functie maakt een connectie met de databank met LSVI-indicatoren.  Deze connectie moet altijd geopend worden en meegegeven als argument om andere functies te kunnen gebruiken, want alle functies binnen dit package hebben een databankconnectie nodig om gegevens uit de databank te halen.  Zonder deze connectie werken ze niet.  Binnen INBO kunnen de defaultwaarden gebruikt worden, d.w.z. dat er geen argumenten meegegeven moeten worden in connecteerMetLSVIdb().  Gebruikers buiten INBO hebben een kopie van de databank nodig om met dit package te kunnen werken, en eventueel een aangepaste versie van deze functie.
}
\examples{
library(RODBC)
ConnectieLSVIhabitats <- connecteerMetLSVIdb()
sqlQuery(ConnectieLSVIhabitats, "SELECT VersieLSVI, Referentie FROM Versie")
odbcClose(ConnectieLSVIhabitats)

}

