% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecteerMetLSVIdb.R
\name{connecteerMetLSVIdb}
\alias{connecteerMetLSVIdb}
\title{Connecteer met de databank met LSVI-indicatoren}
\usage{
connecteerMetLSVIdb(Server = "INBO-SQL07-PRD.inbo.be",
  Databank = "D0122_00_LSVIHabitatTypes", Gebruiker = "pc-eigenaar",
  Wachtwoord = "geen")
}
\arguments{
\item{Server}{de server waarop de databank staat die aangeroepen wordt (standaard "INBO-SQL07-PRD.inbo.be")}

\item{Databank}{de naam van de databank die aangeroepen wordt (standaard "D0122_00_LSVIHabitatTypes")}

\item{Gebruiker}{gebruiker van de databank, standaard "pc-eigenaar" waarbij de login en wachtwoord gebruikt wordt waarmee ingelogd is op de pc, andere opties zijn "lezer" waarbij een gebruiker met leesrechten aangemaakt wordt (dit kan enkel als dit geïnstalleerd is op je pc) of een eigen gebruikersnaam en wachtwoord ingeven}

\item{Wachtwoord}{wachtwoord van de gebruiker van de databank, standaard de tekst "geen", moet niet ingevuld worden als Gebruiker "pc-eigenaar" is}
}
\value{
Deze functie geeft een open odbc-connectie naar de gespecifieerde databank.
}
\description{
Deze functie maakt een connectie met de databank met LSVI-indicatorendatabank, wat nodig is om de functies te kunnen gebruiken.  Voorlopig verwijst deze naar een databank binnen INBO, waardoor gebruikers buiten INBO een kopie van de databank nodig hebben om met dit package te kunnen werken, en eventueel een aangepaste versie van deze functie.  Op termijn zal deze databank geïntegreerd worden in het package, waardoor ze overal zou moeten werken.  (Deze functie is zodanig ingebouwd in de code dat ze niet expliciet opgegeven moet worden door de gebruiker, tenzij deze een andere databank wil opgeven.)
}
\examples{
# deze functie, en dus ook onderstaande code, kan enkel gerund worden als er
# een connectie gelegd kan worden met de SQL Server-databank binnen INBO
\dontrun{
library(DBI)
ConnectieLSVIhabitats <- connecteerMetLSVIdb()
dbGetQuery(ConnectieLSVIhabitats, "SELECT VersieLSVI, Referentie FROM Versie")
dbDisconnect(ConnectieLSVIhabitats)
}

}
