% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecteerMetLSVIdb.R
\name{connecteerMetLSVIdb}
\alias{connecteerMetLSVIdb}
\title{Connecteer met de databank met LSVI-indicatoren}
\usage{

  connecteerMetLSVIdb(path = "C://Users/els.lommelen@inbo.be/els.lommelen@inbo.be/10194_Controlelijsten/Output/LSVIdb_v4.accdb")
}
\arguments{
\item{path}{Het path waarlangs de databank toegankelijk is (inclusief databanknaam)}
}
\value{
Deze functie geeft een open odbc-connectie naar de gespecifieerde databank.
}
\description{
Deze functie maakt een connectie met de databank met LSVI-indicatoren.  Bedoeling van deze aparte functie is om de link naar deze databank maar op 1 plaats te bewaren, zodat deze bij het verplaatsen van de databank ook maar op een plaats aangepast moet worden.  Deze link wordt als argument meegegeven om gebruikers toe te laten om deze aan te passen.  (Deze link zou eigenlijk ook aanpasbaar moeten zijn in de functies die deze aanroepen!)
}
\examples{
library(RODBC)
connectie <- connecteerMetLSVIdb()
sqlQuery(connectie, "SELECT VersieLSVI, Referentie FROM Versie")
odbcClose(connectie)

}

