% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecteerMetLSVIlite.R
\name{connecteerMetLSVIlite}
\alias{connecteerMetLSVIlite}
\title{Connecteer met de databank met LSVI-indicatoren in het package}
\usage{
connecteerMetLSVIlite()
}
\value{
Deze functie geeft een open odbc-connectie naar de SQLite-databank in de installatie-file van het package.
}
\description{
Deze functie maakt een connectie met de in het package toegevoegde databank met LSVI-indicatoren, net zoals de functie connecteerMetLSVIdb().    Aangeraden wordt om de functie connecteerMetLSVIdb() te gebruiken in plaats van deze functie, omdat deze functie op termijn zal verdwijnen.
}
\examples{
library(LSVI)
library(DBI)
ConnectieLSVIhabitats <- connecteerMetLSVIlite()
dbGetQuery(ConnectieLSVIhabitats, "SELECT VersieLSVI, Referentie FROM Versie")
dbDisconnect(ConnectieLSVIhabitats)

}
