% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefInfoHabitatfiche.R
\name{geefInfoHabitatfiche}
\alias{geefInfoHabitatfiche}
\title{Geeft tabel met info uit de LSVI-rapporten voor de opgegeven parameters}
\usage{
geefInfoHabitatfiche(Versie = geefUniekeWaarden("Versie", "VersieLSVI"),
  Habitatgroep = geefUniekeWaarden("Habitatgroep", "Habitatgroepnaam"),
  Habitattype = geefUniekeWaarden("Habitattype", "Habitatcode"),
  Habitatsubtype = geefUniekeWaarden("Habitatsubtype", "Habitatcode_subtype"),
  Criterium = geefUniekeWaarden("Criterium", "Naam"),
  Indicator = geefUniekeWaarden("Indicator", "Naam"), Stijl = c("Rmd",
  "tekst"))
}
\arguments{
\item{Versie}{De versie van het LSVI-rapport, bv. "Versie 2" of "Versie 3".  Bij de default "alle" worden de gegevens voor de verschillende versies gegeven.}

\item{Habitatgroep}{Parameter waarmee alle habitats van een bepaalde habitatgroep kunnen geselecteerd worden, bv. "Bossen", "Heiden", "(Half-)natuurlijke graslanden", "Zoete wateren",...   en "alle" (=default).  Deze waarde moet niet gespecifieerd worden als een bepaald habitat(sub)type geselecteerd wordt.}

\item{Habitattype}{Parameter waarmee een habitattype kan geselecteerd worden.  Als dit habitattype meerdere subtypes heeft, zullen de gegevens van alle subtypes van dit habitattype weergegeven worden.}

\item{Habitatsubtype}{Parameter waarmee een habitatsubtype geselecteerd kan worden.  Als deze waarde ingevuld is, is het niet nodig om de parameters Habitatgroep en Habitattype te specifieren.}

\item{Criterium}{Het LSVI-criterium waarvoor de gegevens geselecteerd worden: "Vegetatie", "Structuur", "Verstoring" of "alle".}

\item{Indicator}{De indicator waarvoor de gegevens uit de databank gehaald worden.}

\item{Stijl}{Keuze uit "Rmd" en "tekst".  Bij Rmd (default) worden soortgroepnamen voorafgegaan en gevolgd door "__" en Latijnse namen van soorten door "_", waardoor deze bij gebruik van RMarkdown worden omgezet naar resp. vet en italics.  Bij tekst worden deze underscores weggelaten.}
}
\value{
Deze functie genereert een tabel met alle gegevens die nodig zijn om de tabellen habitatkarakteristieken en beoordelingsmatrix uit de LSVI-rapporten te genereren.
}
\description{
Deze functie geeft de inhoud van de tabellen habitatkarakteristieken en beoordelingsmatrix uit de rapporten van de Lokale Staat van Instandhouding voor de habitatsubtypes die voldoen aan de opgegeven parameters.  Volledigheidshalve geeft ze ook de uigebreide namen van de habitattypes en habitatsubtypes.

De parameters kunnen enkel de hieronder gespecifeerde waarden bevatten en moeten als string opgegeven worden.  Default is telkens "alle", waarbij de soortenlijsten voor alle mogelijke waarden van die parameter weergegeven worden (m.a.w. er is geen selectie voor deze parameter).
}
\examples{
geefInfoHabitatfiche(Versie = "Versie 3", Habitatsubtype = "4010")

}

