% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefSoortenlijst.R
\name{geefSoortenlijst}
\alias{geefSoortenlijst}
\title{Genereert soortenlijst(en) LSVI op basis van de opgegeven parameters}
\usage{
geefSoortenlijst(
  Versie = "alle",
  Habitatgroep = "alle",
  Habitattype = "alle",
  Criterium = "alle",
  Indicator = "alle",
  Taxonlijstniveau = c("habitattype", "criterium", "indicator", "voorwaarde"),
  Taxonlijsttype = deprecated(),
  ConnectieLSVIhabitats = NULL
)
}
\arguments{
\item{Versie}{De versie van het LSVI-rapport, bv. "Versie 2" of "Versie 3".
Bij de default "alle" worden de gegevens voor de verschillende versies
gegeven.  De mogelijke waarden kunnen opgevraagd worden via
\code{geefUniekeWaarden("Versie", "VersieLSVI")} of \code{geefVersieInfo()}.}

\item{Habitatgroep}{Parameter waarmee alle habitats van een bepaalde
habitatgroep kunnen geselecteerd worden, bv. "Bossen", "Heiden",
"(Half-)natuurlijke graslanden", "Zoete en brakke wateren",...   en "alle"
(=default).  Deze waarde moet niet gespecifieerd worden als een bepaald
habitat(sub)type geselecteerd wordt.  De mogelijke waarden kunnen opgevraagd
worden via \code{geefUniekeWaarden("Habitatgroep", "Naam")}.}

\item{Habitattype}{Parameter waarmee een habitattype of habitatsubtype kan
geselecteerd worden.  Als dit een habitattype betreft met meerdere subtypes,
zullen de gegevens van alle subtypes van dit habitattype weergegeven worden.
De mogelijke waarden kunnen opgevraagd worden via
\code{geefUniekeWaarden("Habitattype", "Code")}.  Er is voor deze parameter ook de
mogelijkheid om een vector van meerdere habitat(sub)typen op te geven.}

\item{Criterium}{Het LSVI-criterium waarvoor de gegevens geselecteerd
worden: "Vegetatie", "Structuur", "Verstoring" of "alle".}

\item{Indicator}{De indicator waarvoor de gegevens uit de databank gehaald
worden.  De mogelijke waarden kunnen opgevraagd worden via
\code{geefUniekeWaarden("Indicator", "Naam")}.}

\item{Taxonlijstniveau}{Geeft aan op welk niveau de soortenlijst gegroepeerd
is (en welke niveaus weergegeven worden in de soortenlijst), de mogelijke
waarden zijn "habitattype", "criterium", "indicator" en "voorwaarde".
Default is "habitattype".}

\item{Taxonlijsttype}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{Taxonlijsttype = "alle"} wordt niet meer ondersteund; deze functie zal
altijd de soortenlijsten weergeven zoals in de habitatfiche (na herziening
van de soortafhandeling bevat het package geen volledige taxonomische
lijsten meer en is die functionaliteit ook overbodig geworden)}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor
de LSVI van habitats, in te stellen d.m.v. functie \code{connecteerMetLSVIdb()}.}
}
\value{
Deze functie geeft een tabel met velden \code{Versie}, \code{Habitattype},
\code{Habitatsubtype}, \code{Criterium}, \code{Indicator}, evt. \code{Beschrijving}, \code{WetNaam},
\code{WetNaamKort} en \code{NedNaam} (waarbij \code{Beschrijving} een omschrijving is voor
een groep van taxa binnen eenzelfde indicator).  \code{WetNaam} is de volledige
Latijnse naam inclusief auteursnaam, \code{WetNaamKort} geeft de verkorte naam
zonder auteursnaam.
Daarnaast heeft de tabel ook de velden \code{GbifUsageKey} (unieke ID van Gbif)
en \code{Rank} (niveau van taxon en \code{GbifUsageKey}) die bij de berekeningen
gebruikt worden om de taxa van de opname te koppelen.
}
\description{
Deze functie genereert soortenlijsten (met wetenschappelijke en
Nederlandse namen) die gebruikt worden voor de bepaling van de Lokale Staat
van Instandhouding van de opgegeven parameters.
In feite genereert ze een tabel met velden \code{Versie}, \code{Habitattype},
\code{Habitatsubtype}, \code{WetNaam}, \code{WetNaamKort} en \code{NedNaam} en evt. \code{Criterium},
\code{Indicator} en/of \code{Beschrijving} waarin de
gespecifieerde parameters uitgeselecteerd zijn en waar voor andere
parameters alle waarden uit de databank weergegeven zijn.

Voor de vorm van de soortenlijst zijn er meerdere opties: een soortenlijst
met alle soorten per habitat(sub)type, ofwel gegroepeerd per criterium,
indicator of voorwaarde.  Dit kan opgegeven worden in de parameter
Taxonlijstniveau.

De parameters kunnen enkel de hieronder gespecifieerde waarden bevatten en moeten als string opgegeven worden.  Default is telkens "alle", waarbij de soortenlijsten voor alle mogelijke waarden van die parameter weergegeven worden (m.a.w. er is geen selectie voor deze parameter).
}
\examples{
# Omwille van de iets langere lange duurtijd van de commando's staat bij
# onderstaande voorbeelden de vermelding 'dontrun' (om problemen te vermijden
# bij het testen van het package). Maar de voorbeelden werken en kunnen zeker
# uitgetest worden.
\dontrun{
maakConnectiePool()
geefSoortenlijst(Habitattype = "4030")
library(pool)
poolClose(ConnectiePool)
}

}
