% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefSoortenlijstVoorIDs.R
\name{geefSoortenlijstVoorIDs}
\alias{geefSoortenlijstVoorIDs}
\title{Genereert soorten(groep)lijst(en) LSVI op basis van TaxongroepID}
\usage{
geefSoortenlijstVoorIDs(
  Taxongroeplijst,
  Taxonlijsttype = c("LSVIfiche", "alle"),
  ConnectieLSVIhabitats = NULL
)
}
\arguments{
\item{Taxongroeplijst}{string waarin de \code{TaxongroepId}'s na elkaar
weergegeven worden, gescheiden door een komma.
Eventueel mag dit ook een vector zijn van \code{TaxongroepId}'s.}

\item{Taxonlijsttype}{\code{"LSVIfiche"} betekent dat de taxonlijst van de
habitatfiche wordt overgenomen, \code{"alle"} betekent dat alle soorten en alle
taxonomische groepen worden weergegeven die volledig in de groepen vallen
die aan de parameters voldoen.}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor
de LSVI van habitats, in te stellen d.m.v. functie \code{connecteerMetLSVIdb()}.}
}
\value{
Deze functie geeft een tabel met velden \code{TaxongroepId}, evt.
\code{Beschrijving}, \code{WetNaam}, \code{WetNaamKort} en \code{NedNaam} (waarbij \code{Beschrijving}
een omschrijving is voor een groep van taxons binnen eenzelfde indicator).
\code{WetNaam} is de volledige Latijnse naam inclusief auteursnaam, \code{WetNaamKort}
geeft de verkorte naam zonder auteursnaam.
}
\description{
Deze functie genereert soortenlijsten (met wetenschappelijke en
Nederlandse namen) uit de databank met de criteria en indicatoren voor de
bepaling van de Lokale Staat van Instandhouding.  Het is in feite een
hulpfunctie die voor verschillende andere functies gebruikt wordt en die de
complexe zoekfunctie in de tabellen met soorten uitvoert op basis van een
opgegeven \code{TaxongroepId} (en in die zin iets minder gebruiksvriendelijk is).
Voor een selectie van soortenlijsten op basis van specifieke parameters is
de functie \code{geefSoortenlijst()} een beter alternatief.

Deze functie geeft standaard voor de gespecifieerde taxongroepen per groep
een lijst van alle taxa zoals ze in de LSVI-habitatfiche vermeld zijn
(genusniveau, soortniveau, ondersoort,...).  Op basis van de parameter
soortenlijsttype kan ook gekozen worden om een volledige lijst te geven van
deze taxa en alle taxa die hieronder vallen (en opgenomen zijn in de
onderliggende databank).
}
\examples{
# Omwille van de iets langere lange duurtijd van de commando's staat bij
# onderstaande voorbeelden de vermelding 'dontrun' (om problemen te vermijden
# bij het testen van het package). Maar de voorbeelden werken en kunnen zeker
# uitgetest worden.
\dontrun{
maakConnectiePool()
geefSoortenlijstVoorIDs("434,88,565")
geefSoortenlijstVoorIDs("434,88,565","alle")
library(pool)
poolClose(ConnectiePool)
}

}
