% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefSoortenlijstVoorIDs.R
\name{geefSoortenlijstVoorIDs}
\alias{geefSoortenlijstVoorIDs}
\title{Genereert soorten(groep)lijst(en) LSVI op basis van TaxongroepID}
\usage{
geefSoortenlijstVoorIDs(
  Taxongroeplijst,
  Taxonlijsttype = deprecated(),
  ConnectieLSVIhabitats = NULL
)
}
\arguments{
\item{Taxongroeplijst}{string waarin de TaxongroepId's na elkaar weergegeven
worden, gescheiden door een komma.  Eventueel mag dit ook een vector zijn
van TaxongroepId's.}

\item{Taxonlijsttype}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{Taxonlijsttype = "alle"} wordt niet meer ondersteund; deze functie zal
altijd de soortenlijsten weergeven zoals in de habitatfiche (na herziening
van de soortafhandeling bevat het package geen volledige taxonomische
lijsten meer en is die functionaliteit ook overbodig geworden)}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor
de LSVI van habitats, in te stellen d.m.v. functie connecteerMetLSVIdb.}
}
\value{
Deze functie geeft een tabel met velden TaxongroepId, evt.
Beschrijving, WetNaam, WetNaamKort en NedNaam (waarbij Beschrijving een
omschrijving is voor een groep van taxons binnen eenzelfde indicator).
WetNaam is de volledige Latijnse naam inclusief auteursnaam, WetNaamKort
geeft de verkorte naam zonder auteursnaam.
}
\description{
Deze functie genereert soortenlijsten (met wetenschappelijke en
Nederlandse namen) uit de databank met de criteria en indicatoren voor de
bepaling van de Lokale Staat van Instandhouding.  Het is in feite een
hulpfunctie die voor verschillende andere functies gebruikt wordt en die de
complexe zoekfunctie in de tabellen met soorten uitvoert op basis van een
opgegeven TaxongroepId (en in die zin iets minder gebruiksvriendelijk is).
Voor een selectie van soortenlijsten op basis van specifieke parameters is
de functie geefSoortenlijst() een beter alternatief.

Deze functie geeft voor de gespecifieerde taxongroepen per groep
een lijst van alle taxa zoals ze in de LSVI-habitatfiche vermeld zijn
(genusniveau, soortniveau, subsoort,...) en ook de rank van de taxa.
}
\examples{
# Omwille van de iets langere lange duurtijd van de commando's staat bij
# onderstaande voorbeelden de vermelding 'dontrun' (om problemen te vermijden
# bij het testen van het package). Maar de voorbeelden werken en kunnen zeker
# uitgetest worden.
\dontrun{
maakConnectiePool()
geefSoortenlijstVoorIDs("434,88,565")
library(pool)
poolClose(ConnectiePool)
}

}
