% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefSoortenlijstVoorIDs.R
\name{geefSoortenlijstVoorIDs}
\alias{geefSoortenlijstVoorIDs}
\title{Genereert soorten(groep)lijst(en) LSVI op basis van SoortengroepID}
\usage{
geefSoortenlijstVoorIDs(Soortengroeplijst, Soortenlijsttype = c("LSVIfiche",
  "alle"), ConnectieLSVIhabitats = ConnectiePool)
}
\arguments{
\item{Soortengroeplijst}{string waarin de SoortengroepID's na elkaar weergegeven worden, gescheiden door een komma.  Eventueel mag dit ook een vector zijn van SoortengroepID's.}

\item{Soortenlijsttype}{"LSVIfiche" betekent dat de soortenlijst van de habitatfiche wordt overgenomen, "alle" betekent dat alle soorten en alle taxonomische groepen worden weergegeven die volledig in de groepen vallen die aan de parameters voldoen.}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor de LSVI van habitats, in te stellen d.m.v. functie connecteerMetLSVIdb.}
}
\value{
Deze functie geeft een tabel met velden SoortengroepID, evt. Beschrijving, WetNaam, WetNaamKort en NedNaam (waarbij Beschrijving een omschrijving is voor een groep van soorten binnen eenzelfde indicator).  WetNaam is de volledige Latijnse naam inclusief auteursnaam, WetNaamKort bevat enkel genusnaam en soortnaam (zonder auteursnaam).
}
\description{
Deze functie genereert soortenlijsten (met wetenschappelijke en Nederlandse namen) uit de databank met de criteria en indicatoren voor de bepaling van de Lokale Staat van Instandhouding.  Het is in feite een hulpfunctie die voor verschillende andere functies gebruikt wordt en die de complexe zoekfunctie in de tabellen met soorten uitvoert op basis van een opgegeven SoortengroepID (en in die zin iets minder gebruiksvriendelijk is).  Voor een selectie van soortenlijsten op basis van specifieke parameters is de functie geefSoortenlijst() een beter alternatief.

Deze functie geeft standaard voor de gespecifieerde soortengroepen per soortengroep een lijst van alle taxa zoals ze in de LSVI-habitatfiche vermeld zijn (genusniveau, soortniveau, subsoort,...).  Op basis van de parameter soortenlijsttype kan ook gekozen worden om een volledige lijst te geven van deze taxa en alle taxa die hieronder vallen (en opgenomen zijn in de onderliggende databank).
}
\examples{
geefSoortenlijstVoorIDs("98,227,484,552,726")
geefSoortenlijstVoorIDs("98,227,484,552,726","alle")

}
