% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoercontroleVersie.R
\name{invoercontroleVersie}
\alias{invoercontroleVersie}
\title{Invoercontrole voor waarde versie}
\usage{
invoercontroleVersie(Versie, ConnectieLSVIhabitats)
}
\arguments{
\item{Versie}{Waarde waarop invoercontrole moet gebeuren.}

\item{ConnectieLSVIhabitats}{Connectie met de databank met indicatoren voor de LSVI van habitats, in te stellen d.m.v. functie connecteerMetLSVIdb.}
}
\description{
Om te vermijden dat we meermaals dezelfde invoercontrole moeten uitvoeren en om de hoofdscripts overzichtelijk te houden, maken we voor elke invoercontrole een aparte hulpfunctie aan, die we kunnen aanroepen.  Deze wordt NIET geëxporteerd, dus deze functies kunnen niet als commando gerund worden (maar worden wel gerund als de functie waarin ze voorkomen, aangeroepen wordt).
}
