% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maakConnectiepoolSQLite.R
\name{maakConnectiepoolSQLite}
\alias{maakConnectiepoolSQLite}
\title{Maak een connectiepool met de databank met LSVI-indicatoren in het package}
\usage{
maakConnectiepoolSQLite()
}
\value{
Deze functie maakt een Environment-object aan dat de connecties regelt met de betreffende databank in de installatie-file van het package.
}
\description{
Deze functie maakt een connectiepool met de in het package toegevoegde databank met LSVI-indicatoren, net zoals de functie maakConnectiePool().  Aangeraden wordt om de functie maakConnectiePool() te gebruiken in plaats van deze functie, omdat deze functie op termijn zal verdwijnen.
}
\examples{
# Omwille van de iets langere lange duurtijd van de commando's staat bij
# onderstaand voorbeeld de vermelding 'dontrun' (om problemen te vermijden
# bij het testen van het package). Maar het voorbeeld werkt en kan zeker
# uitgetest worden.
\dontrun{
library(LSVI)
maakConnectiepoolSQLite()
geefVersieInfo()
library(pool)
poolClose(ConnectiePool)
}

}
