% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maakHabitatfiches.R
\name{maakHabitatfiches}
\alias{maakHabitatfiches}
\title{Genereert habitatfiche(s) van LSVI op basis van de opgegeven parameters}
\usage{
maakHabitatfiches(Versie = geefUniekeWaarden("Versie", "VersieLSVI"),
  Habitatgroep = geefUniekeWaarden("Habitatgroep", "Habitatgroepnaam"),
  Habitattype = geefUniekeWaarden("Habitattype", "Habitatcode"),
  Habitatsubtype = geefUniekeWaarden("Habitatsubtype", "Habitatcode_subtype"),
  verbose = TRUE, Server = "inbosql03\\\\prd",
  Databank = "D0122_00_LSVIHabitatTypes", Gebruiker = "D0122_AppR",
  Wachtwoord = "19D939F1-BCCE-439F-9ED4-6A886E038A6D")
}
\arguments{
\item{Versie}{De versie van het LSVI-rapport, bv. "Versie 2" of "Versie 3".  Bij de default "alle" worden de soortenlijsten voor de verschillende versies gegeven.}

\item{Habitatgroep}{Parameter waarmee alle habitats van een bepaalde habitatgroep kunnen geselecteerd worden, bv. "Bossen", "Heiden", "(Half-)natuurlijke graslanden", "Zoete wateren",...   en "alle" (=default).  Deze waarde moet niet gespecifieerd worden als een bepaald habitat(sub)type geselecteerd wordt.}

\item{Habitattype}{Parameter waarmee een habitattype kan geselecteerd worden.  Als dit habitattype meerdere subtypes heeft, zullen de soortenlijsten van alle subtypes van dit habitattype weergegeven worden.}

\item{Habitatsubtype}{Parameter waarmee een habitatsubtype geselecteerd kan worden.  Als deze waarde ingevuld is, is het niet nodig om de parameters Habitatgroep en Habitattype te specifieren.}

\item{verbose}{geeft de toestand van het systeem aan, om te zorgen dat boodschappen niet onnodig gegeven worden}

\item{Server}{de server waarop de databank staat die aangeroepen wordt (standaard "inbosql03\\prd")}

\item{Databank}{de naam van de databank die aangeroepen wordt (standaard "D0122_00_LSVIHabitatTypes")}

\item{Gebruiker}{gebruiker van de databank, standaard een gebruiker met leesrechten, andere opties zijn "pc-eigenaar" waarbij de login en wachtwoord gebruikt wordt waarmee ingelogd is op de pc of een eigen gebruikersnaam en wachtwoord ingeven}

\item{Wachtwoord}{wachtwoord van de gebruiker van de databank, standaard een gebruiker met leesrechten, moet niet ingevuld worden als Gebruiker "pc-eigenaar" is}
}
\value{
Deze functie genereert habitatfiches in de vorm van html-files die in de workspace opgeslagen worden.
}
\description{
Deze functie genereert habitatfiches die gebruikt worden voor de bepaling van de Lokale Staat van Instandhouding van de habitatsubtypes die voldoen aan de opgegeven parameters.  

De parameters kunnen enkel de hieronder gespecifeerde waarden bevatten en moeten als string opgegeven worden.  Default is telkens "alle", waarbij de soortenlijsten voor alle mogelijke waarden van die parameter weergegeven worden (m.a.w. er is geen selectie voor deze parameter).

De gegenereerde habitatfiches worden opgeslagen in de folder die als working directory gespecifieerd is.
}
\examples{
maakHabitatfiches(Versie = "Versie 3", Habitatsubtype = "4010")

}

