% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTaxonnaam.R
\name{parseTaxonnaam}
\alias{parseTaxonnaam}
\title{Vereenvoudig de soortnaam door de auteurs te verwijderen}
\usage{
parseTaxonnaam(Taxonnaam, ParseType = "canonicalnamewithmarker")
}
\arguments{
\item{Taxonnaam}{Wetenschappelijke naam of namen die vereenvoudigd moeten
worden (String)}

\item{ParseType}{Welk type parsing teruggegeven moet worden.  Standaard is
dit \code{"canonicalnamewithmarker"}, andere opties zijn \code{"canonicalname"} en
\code{"canonicalnamecomplete"}}
}
\value{
Deze functie geeft de vereenvoudigde naam of namen terug (als string
of vector van strings)
}
\description{
Deze functie vereenvoudigt de opgegeven taxonna(a)m(en) door
de auteursnaam te verwijderen.  Ze is gebaseerd op de functie name_parse
uit het package rgbif, maar ze vangt enkele specifieke situaties uit de
INBO-plantenlijsten wel op die de originele functie niet correct opvangt, bv.
v., an en den herkennen als deel van een auteursnaam en s.l. en meerdere
soortnamen gescheiden door / (zonder spatie) vervangen door een achtervoegsel
groep.
}
\examples{
parseTaxonnaam("Rosa canina L.")

}
