% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTaxonnaam.R
\name{parseTaxonnaam}
\alias{parseTaxonnaam}
\title{Vereenvoudig de soortnaam door de auteurs te verwijderen}
\usage{
parseTaxonnaam(Taxonnaam, ParseType = "canonicalnamewithmarker")
}
\arguments{
\item{Taxonnaam}{Wetenschappelijke naam of namen die vereenvoudigd moeten
worden (String)}

\item{ParseType}{Welk type parsing teruggegeven moet worden.  Standaard is
dit \code{"canonicalnamewithmarker"}, andere opties zijn \code{"canonicalname"} en
\code{"canonicalnamecomplete"}}
}
\value{
Deze functie geeft de vereenvoudigde naam of namen terug (als string
of vector van strings)
}
\description{
Deze functie vereenvoudigt de opgegeven taxonnamen door
de auteursnaam te verwijderen.  Ze is gebaseerd op de functie \code{name_parse()}
uit het package \code{rgbif}, maar ze vangt enkele specifieke situaties uit de
INBO-plantenlijsten wel op die de originele functie niet correct opvangt, bv.
\code{v.}, \code{an} en \code{den} herkennen als deel van een auteursnaam en
\code{s.l.} en meerdere soortnamen gescheiden door \code{/} (zonder spatie) vervangen
door een achtervoegsel \code{groep}.
}
\examples{
parseTaxonnaam("Rosa canina L.")

}
