% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertaalInvoerInterval.R
\name{vertaalInvoerInterval}
\alias{vertaalInvoerInterval}
\title{zet ingevoerde gegevens om naar een interval}
\usage{
vertaalInvoerInterval(Dataset, LIJST)
}
\arguments{
\item{Dataset}{dataframe met velden Rijnr, Type, Waarde, Eenheid en Invoertype}
}
\value{
Dataframe met velden Min
}
\description{
Deze functie zet ingevoerde gegevens van meerdere types om naar een interval bestaande uit minimumwaarde en maximumwaarde.  De functie gebruikt Type, Eenheid en Invoertype om te bepalen welke omzetting eventueel nodig is.  Percentages worden bv. omgezet naar een decimaal getal (waarbij minimum en maximum dezelfde waarde zullen krijgen), en categorische variabelen met Invoertype Tansley worden omgezet naar de onder- en bovengrens die in de databank gegeven worden voor de betreffende categorie.
}
\examples{


}
